pull: Apply -q and -v options to rebase mode as well
git pull passed -q and -v only to git merge, but they can be useful for
git rebase as well, so pass them there, too.

In particular, using -q shuts up the "Already up-to-date." message.
Especially, a new test script runs the same "pull --rebase" twice to
make sure both cases are quiet, when it has something to fetch and
when it is already up to date.

Signed-off-by: Peter Eisentraut <peter@eisentraut.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6288b6f19d83ff0987b8be81fc92f99499a9599b [formerly 8b5c827658f10a924dadd8079a3bac4a0f4a21e6] [formerly 8b5c827658f10a924dadd8079a3bac4a0f4a21e6 [formerly ce4c4d4ec3df274349f20431c61ed85cea038026]]
Former-commit-id: 862f63952f9b6088557cca2ac690bc6ea18bac74
Former-commit-id: 87551aeecd46f9870c749152dc111d7fae4695ed