archive-zip: use deflateInit2() to ask for raw compressed data
We use the function git_deflate_init() -- which wraps the zlib function
deflateInit() -- to initialize compression of ZIP file entries.  This
results in compressed data prefixed with a two-bytes long header and
followed by a four-bytes trailer.  ZIP file entries consist of ZIP
headers and raw compressed data instead, so we remove the zlib wrapper
before writing the result.

We can ask zlib for the the raw compressed data without the unwanted
parts in the first place by using deflateInit2() and specifying a
negative number of bits to size the window.  For that purpose, factor
out the function do_git_deflate_init() and add git_deflate_init_raw(),
which wraps it.  Then use the latter in archive-zip.c and get rid of
the code that stripped the zlib header and trailer.

Also rename the helper function zlib_deflate() to zlib_deflate_raw()
to reflect the change.

Thus we avoid generating data that we throw away anyway, the code
becomes shorter and some magic constants are removed.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c456543d7f3c20c9181b807c319230ddcda33fbf [formerly f753ff207e1a1f29b433700788bbe6b2646fa9d2] [formerly f753ff207e1a1f29b433700788bbe6b2646fa9d2 [formerly c3c2e1a09b84dc932d90fa09d1632e5efa34c940]]
Former-commit-id: 3ae7047f24e7973609f52d7f236f238873ac158d
Former-commit-id: dee41f0ea29d0c339eb58dca5b95c1ab91eff479