wt-status: move strbuf into read_and_strip_branch()
The strbufs are placed outside read_and_strip_branch as a premature
optimization: when it reads "refs/heads/foo" to strbuf and wants to
return just "foo", it could do so without memory movement. In return
the caller must not use the returned pointer after releasing strbufs,
which own the buffers that contain the returned strings. It's a clumsy
design.

By moving strbufs into read_and_strip_branch(), the returned pointer
always points to a malloc'd buffer or NULL. The pointer can be passed
around and freed after use.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3bee674c3616c8bee06875f2335b42da50a75e18 [formerly 5fa7e9bc97ae9da1894ec1a6a5a626004d6941d1] [formerly 5fa7e9bc97ae9da1894ec1a6a5a626004d6941d1 [formerly 8b87cfd000c7c98f58279cff698f6e8c7892c059]]
Former-commit-id: f09d9c97c461178e6c278a66282243ba06b07cb4
Former-commit-id: 8d382bd03eeddfb75efe5fc0432d69fff8bd6986