rev-parse: clarify documentation of $name@{upstream} syntax
"git rev-parse" interprets string in string@{upstream} as a name of
a branch not a ref. For example, refs/heads/master@{upstream} looks
for an upstream branch that is merged by git-pull to ref
refs/heads/refs/heads/master not to refs/heads/master.

However the documentation could mislead a user to believe that the
string is interpreted as ref.

Signed-off-by: Kacper Kornet <draenog@pld-linux.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5fcbadeca18e706db016119eb2499ad3edc21191 [formerly 7239c9d8ed5739973b5ca99380e097662cbf5420] [formerly 7239c9d8ed5739973b5ca99380e097662cbf5420 [formerly 47e329ef7c46de8ccdc8c4e40333579264a83c21]]
Former-commit-id: 94fd464fb5b89b14a5c62dbf3ec429949db881a5
Former-commit-id: b917b7950a1071c6574687bd983e9a88cc360b55