avoid segfaults on parse_object failure
Many call-sites of parse_object assume that they will get a
non-NULL return value; this is not the case if we encounter
an error while parsing the object.

This patch adds a wrapper function around parse_object that
handles dying automatically, and uses it anywhere we
immediately try to access the return value as a non-NULL
pointer (i.e., anywhere that we would currently segfault).

This wrapper may also be useful in other places. The most
obvious one is code like:

  o = parse_object(sha1);
  if (!o)
	  die(...);

However, these should not be mechanically converted to
parse_object_or_die, as the die message is sometimes
customized. Later patches can address these sites on a
case-by-case basis.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 93e1a4c572ee5efb96e6fb30eacea29b4d9af582 [formerly 93e1a4c572ee5efb96e6fb30eacea29b4d9af582 [formerly 75a95490474ab6e991cbbbd10d980498a9109648]]
Former-commit-id: 21cab3b09be2c97dcce5fbb47521b8db669f387c
Former-commit-id: b2883378ce63c74e2b69047d047dda8842c60d47