use parse_object_or_die instead of die("bad object")
Some call-sites do:

  o = parse_object(sha1);
  if (!o)
	  die("bad object %s", some_name);

We can now handle that as a one-liner, and get more
consistent output.

In the third case of this patch, it looks like we are losing
information, as the existing message also outputs the sha1
hex; however, parse_object will already have written a more
specific complaint about the sha1, so there is no point in
repeating it here.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 322f81394229603c2f5c80f40db77d06595bb330 [formerly 109113c27c3cbacf5d8a8721629caa8dfd1c8245] [formerly 109113c27c3cbacf5d8a8721629caa8dfd1c8245 [formerly f7892d181752187513f10b13f2272fa46c9c8422]]
Former-commit-id: 0f76758e6716c78541c6b3a6d0429eccdff5469d
Former-commit-id: dd6cee3baed369e6729e25cf656835d9f620eaae