safe_create_leading_directories: fix race that could give a false negative
If two processes are racing to create the same directory tree, they
will both see that the directory doesn't exist, both try to mkdir(),
and one of them will fail.  This is okay, as we only care that the
directory gets created.  So, we add a check for EEXIST from mkdir,
and continue when the directory exists, taking the same codepath as
the case where the earlier stat() succeeds and finds a directory.

Signed-off-by: Steven Walter <stevenrwalter@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a381436cb81da6a53238f52e0fccfe395be10f0a [formerly 548285f8577e780a83c75d6cbbda096142b72cf2] [formerly 548285f8577e780a83c75d6cbbda096142b72cf2 [formerly 928734d99370e2478e3b3881e5b0aa79745ac27c]]
Former-commit-id: c2bf2203fd3875272a208a5c15889631397c387a
Former-commit-id: 374bfa83db92925d37b4af63a05d2baf6b69ba24