push test: use test_config when appropriate
Configuration from test_config does not last beyond the end of the
current test assertion, making each test easier to think about in
isolation.

This changes the meaning of some of the tests.  For example, currently
"push with insteadOf" passes even if the line setting
"url.$TRASH.pushInsteadOf" is dropped because an url.$TRASH.insteadOf
setting leaks in from a previous test.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 56fd448eec213ac80ce3d64ae7d00ffd1ff8cf8e [formerly 403c0c44bde5c4a9fee771f7ea304f3815d31b81] [formerly 403c0c44bde5c4a9fee771f7ea304f3815d31b81 [formerly 3c695523384855dffba49aec7f0a77e555ced822]]
Former-commit-id: b32bb05f2749f5d4f684c33ed98034080249fb5d
Former-commit-id: c09e19dc6e972b927572f7957f4a708f585af861