push test: simplify check of push result
This test checks each ref with code like the following:

	r=$(git show-ref -s --verify refs/$ref) &&
	test "z$r" = "z$the_first_commit"

Afterward it counts refs:

	test 1 = $(git for-each-ref refs/remotes/origin | wc -l)

Simpler to test the number and values of relevant refs in for-each-ref
output at the same time using test_cmp.  This makes the test more
readable and provides more helpful "./t5516-push-push.sh -v" output
when the test fails.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bc0d126abd032aab85c1e668c2f038168b212321 [formerly bc0d126abd032aab85c1e668c2f038168b212321 [formerly 848575d833ca0d9d47db372348576f3b98bd19b7]]
Former-commit-id: 22ed2c56689a6543ed2368b28f1966c33245a5ce
Former-commit-id: 9d96329503910541bf522da500260c039344e69b