index-pack: protect deepest_delta in multithread code
deepest_delta is a global variable but is updated without protection
in resolve_delta(), a multithreaded function. Add a new mutex for it,
but only protect and update when it's actually used (i.e. show_stat is
non-zero).

Another variable that will not be updated is delta_depth in "struct
object_entry" as it's only useful when show_stat is 1. Putting it in
"if (show_stat)" makes it clearer.

The local variable "stat" is renamed to "show_stat" after moving to
global scope because the name "stat" conflicts with stat(2) syscall.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 64e9ac19665bada5e9241d0c17af8e3edd19e70a [formerly 64e9ac19665bada5e9241d0c17af8e3edd19e70a [formerly 3aba2fddcbff6ba611c05eabeb4a8561c119275f]]
Former-commit-id: c21601fd0feeb6f47d9183715fefd17401be1751
Former-commit-id: 0241106d5c4d76a7d5908ee428f4888b11a9bb2e