index-pack: guard nr_resolved_deltas reads by lock
The threaded parts of index-pack increment the number of resolved
deltas in nr_resolved_deltas guarded by counter_mutex.  However, the
per-thread outer loop accessed nr_resolved_deltas without any locks.

This is not wrong as such, since it doesn't matter all that much
whether we get an outdated value.  However, unless someone proves that
this one lock makes all the performance difference, it would be much
cleaner to guard _all_ accesses to the variable with the lock.

The only such use is display_progress() in the threaded section (all
others are in the conclude_pack() callchain outside the threaded
part).  To make it obvious that it cannot deadlock, move it out of
work_mutex.

Signed-off-by: Thomas Rast <trast@student.ethz.ch>
Reviewed-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0e73a8dab5760eca1aaf08bcad364970d8b912bd [formerly 30fee41fe4e97d1e1a9c3f458c94c5031daa93dc] [formerly 30fee41fe4e97d1e1a9c3f458c94c5031daa93dc [formerly 8f82aad4e7baf32942eb2087a882e0f8646b2da8]]
Former-commit-id: f7ded5ec352c3b932d7365445fb23ac5e0462882
Former-commit-id: 959601738e7aea99e467a8a491f727bd7031d6c8