add: make pathless 'add [-u|-A]' warning a file-global function
Currently warn_pathless_add() is only called directly by cmd_add(),
but that is about to change.  Move its definition higher in the file
and pass the "--update" or "--all" option name used in its message
through globals instead of function arguments to make it easier to
call without passing values that will not change through the call
chain.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 580c1c8975e7f65b989a3af4b7147e90f3ec7029 [formerly 580c1c8975e7f65b989a3af4b7147e90f3ec7029 [formerly e24afab09137d785f07e62aee8936b548436d21f]]
Former-commit-id: 4d2526224124edf6a9f2c9708a4d9b2ccc09b0dc
Former-commit-id: d22b224f78d87cdea83b366c17ecd3b0aec5be36