git-am: show the final log message on "Applying:" indicator
The "Applying:" message "git am" shows to tell the user which patch
is being applied has traditionally been to help identifying the
input, but we started showing the edited result since f23272f3fd84
(git-am -i: report rewritten title, 2007-12-04), because it was
found more confusing to show the original during an interactive
session.

Treat the modification by the applypatch-msg hook in a similar way
and use the edited result in the progress indication, even though
this is usually not interactive.

Signed-off-by: Simon Ruderich <simon@ruderich.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5375b08c52be0abfb16bf9517d8a65f9ba66c056 [formerly 5375b08c52be0abfb16bf9517d8a65f9ba66c056 [formerly b34a912989624de37f859b4d70d9eab4cfe285dd]]
Former-commit-id: 85a89725c3b3004710a7c477430f742c66ed0476
Former-commit-id: c47e2e43985b889392f3250c3c7bc713b93b8eb6