drop some obsolete "x = x" compiler warning hacks
In cases where the setting and access of a variable are
protected by the same conditional flag, older versions of
gcc would generate a "might be used unitialized" warning. We
silence the warning by initializing the variable to itself,
a hack that gcc recognizes.

Modern versions of gcc are smart enough to get this right,
going back to at least version 4.3.5. gcc 4.1 does get it
wrong in both cases, but is sufficiently old that we
probably don't need to care about it anymore.

Signed-off-by: Jeff King <peff@peff.net>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: de19e5b917c4fc310d21536bf46cb774e6e47b7b [formerly 59c0b4200a3c98e13663cdbe28bfa9b37eed9e40] [formerly 59c0b4200a3c98e13663cdbe28bfa9b37eed9e40 [formerly cbfd5e1cbb539b82a34195efd0edde20d45a6439]]
Former-commit-id: 09303c25da908bc434fd719aa67942fb7a7ccd49
Former-commit-id: e228292400e63cc827db4b0efea173b1a40be66b