fast-import: clarify "inline" logic in file_change_m
When we read a fast-import line like:

  M 100644 :1 foo.c

we point the local object_entry variable "oe" to the object
named by the mark ":1". When the input uses the "inline"
construct, however, we do not have such an object_entry.

The current code is careful not to access "oe" in the inline
case, but we can make the assumption even more obvious (and
catch violations of it) by setting oe to NULL and adding a
comment. As a bonus, this also squelches an over-zealous gcc
-Wuninitialized warning, which means we can drop the "oe =
oe" initialization hack.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dca46e16fb016ef868d65c34457f1e16a90a024b [formerly cf0be94490b194dfdbc472d1d076d06c1ff036c4] [formerly cf0be94490b194dfdbc472d1d076d06c1ff036c4 [formerly 3aa99df802020f73981d1169c05375d5e2b49cc9]]
Former-commit-id: e5f3877e57edc5604e1f801149672ecf87e478c5
Former-commit-id: 7b10b34d9d326084cff41a69532f2c96072fc0aa