run-command: always set failed_errno in start_command
When we fail to fork, we set the failed_errno variable to
the value of errno so it is not clobbered by later syscalls.
However, we do so in a conditional, and it is hard to see
later under what conditions the variable has a valid value.

Instead of setting it only when fork fails, let's just
always set it after forking. This is more obvious for human
readers (as we are no longer setting it as a side effect of
a strerror call), and it is more obvious to gcc, which no
longer generates a spurious -Wuninitialized warning. It also
happens to match what the WIN32 half of the #ifdef does.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c9c7cff47c4cdc9ccc1525da36d250114f9f580a [formerly c9c7cff47c4cdc9ccc1525da36d250114f9f580a [formerly 25043d8aea7859497c12cb035e6688f76e32ac13]]
Former-commit-id: b2b3749aefde0ab5cf3361b0d1ebe95878f9f2ca
Former-commit-id: 1d5ef3f7de4bfde2581c9c69782dd839690c83a5