Documentation: merging a tag is a special case
When asking Git to merge a tag (such as a signed tag or annotated tag),
it will always create a merge commit even if fast-forward was possible.
It's like having --no-ff present on the command line.

It's a difference from the default behavior described in git-merge.txt.
It should be documented as an exception of "FAST-FORWARD MERGE" section
and "--ff" option description.

Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Reviewed-by: Yann Droneaud <ydroneaud@opteya.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 50d7c7e2e70704c33dcd603cca4d0118a2460266 [formerly 50d7c7e2e70704c33dcd603cca4d0118a2460266 [formerly 77c72780edfa5e3858423366fc91699724793b39]]
Former-commit-id: 17f7cf7b788cbd27b24b76be643dcb79dba7a117
Former-commit-id: 93efa5a807317a4a233c86900562bf7ba916e6d7