apply --whitespace=fix: avoid running over the postimage buffer
Originally update-pre-post-images could assume that any whitespace
fixing will make the result only shorter by unexpanding runs of
leading SPs into HTs and removing trailing whitespaces at the end of
lines.  Updating the post-image we read from the patch to match the
actual result can be performed in-place under this assumption.
These days, however, we have tab-in-indent (aka Python) rule whose
result can be longer than the original, and we do need to allocate
a larger buffer than the input and replace the result.

Fortunately the support for lengthening rewrite was already added
when we began supporting "match while ignoring whitespace
differences" mode in 86c91f91794c (git apply: option to ignore
whitespace differences, 2009-08-04).  We only need to correctly
count the number of bytes necessary to hold the updated result and
tell the function to allocate a new buffer.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ecc2a7f2cdb28846ea3528d7c16093849cc4cd6b [formerly ecc2a7f2cdb28846ea3528d7c16093849cc4cd6b [formerly 250b3c6c992b3cb04e756eb33bed99442fc55193]]
Former-commit-id: 92a6d568cccd752ccbcdd3fe08edc173c23e7609
Former-commit-id: 32eb0a90791d139363b8d5c51cccdee64c666d41