Fix revision walk for commits with the same dates
Logic in still_interesting function allows to stop the commits
traversing if the oldest processed commit is not older then the
youngest commit on the list to process and the list contains only
commits marked as not interesting ones. It can be premature when dealing
with a set of coequal commits. For example git rev-list A^! --not B
provides wrong answer if all commits in the range A..B had the same
commit time and there are more then 7 of them.

To fix this problem the relevant part of the logic in still_interesting
is changed to: the walk can be stopped if the oldest processed commit is
younger then the youngest commit on the list to processed.

Signed-off-by: Kacper Kornet <draenog@pld-linux.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dccdcf7ed03339c4ccb4542a4ebde49db25c81aa [formerly dccdcf7ed03339c4ccb4542a4ebde49db25c81aa [formerly c19d1b4e840535c5fc27077194e8ac219c02644c]]
Former-commit-id: 7ad27fdcebb1886f176258102464b16e0bf5e83a
Former-commit-id: 394de7f62048e8d7418306e7c499c750b535b59c