combine-diff: coalesce lost lines optimally
This replaces the greedy implementation to coalesce lost lines by using
dynamic programming to find the Longest Common Subsequence.

The O(n²) time complexity is obviously bigger than previous
implementation but it can produce shorter diff results (and most likely
easier to read).

List of lost lines is now doubly-linked because we reverse-read it when
reading the direction matrix.

Signed-off-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 809c3b442221c436c4216964ef37629f83449902 [formerly 809c3b442221c436c4216964ef37629f83449902 [formerly 99d3206010ba1fcc9311cbe8376c0b5e78f4a136]]
Former-commit-id: 4bfacd33537c36c8db0d83ec465d2efc09011fa2
Former-commit-id: 0b298ae3d1ef0f20e39484d982a1e49acd6e3c9e