t4202: use test_config/test_unconfig to set/unset git config variables
Instead of using construct such as:
    test_when_finished "git config --unset <key>"
    git config <key> <value>
uses
    test_config <key> <value>
The latter takes care of removing <key> at the end of the test.

Additionally, instead of
     git config <key> ""
or
     git config --unset <key>
uses
     test_unconfig <key>
The latter doesn't failed if <key> is not defined.

Tests are modified to assume correct (default) configuration at entry,
and to reset the modified configuration variables at the end.

Signed-off-by: Yann Droneaud <ydroneaud@opteya.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1b06d9ccf3f4bc9be03414d30ee507fb155dcf12 [formerly 1b06d9ccf3f4bc9be03414d30ee507fb155dcf12 [formerly 90e76b7029d02314b6f8f8470788874fc44fa621]]
Former-commit-id: c215fa0f737123a1a04ad228276296167c21b18a
Former-commit-id: 82a3cca79d3c20f9a90e9481d2901db6bff20be3