t7600: use test_config to set/unset git config variables
Instead of using construct such as:
    test_when_finished "git config --unset <key>"
    git config <key> <value>
uses
    test_config <key> <value>
The latter takes care of removing <key> at the end of the test.

Tests are modified to assume default configuration at entry,
and to reset the modified configuration variables at the end.

Test 'merge log message' was relying on the presence of option `--no-ff`
in the configuration. With the option, git show -s --pretty=format:%b HEAD
produces an empty line and without the option, it produces an empty file.
The test is modified to check with and without `--no-ff` option.

Signed-off-by: Yann Droneaud <ydroneaud@opteya.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: eae4e9847e8b5d78d67d007397bb24880651acc2 [formerly eae4e9847e8b5d78d67d007397bb24880651acc2 [formerly cee683b72eac1a1ac0fcbaabf5256a5f0aff07f7]]
Former-commit-id: 4e80d16b067da6f9840fb736ae4c14926acd79f6
Former-commit-id: c58ae51fcdee73b8486dc950f13d26d1c45472bc