match-trees: simplify score_trees() using tree_entry()
Convert the loop in score_trees() to tree_entry().  The code becomes
shorter and simpler because the calls to update_tree_entry() are not
needed any more.

Another benefit is that we need less variables to track the current
tree entries; as a side-effect of that the compiler has an easier
job figuring out the control flow and thus can avoid false warnings
about uninitialized variables.

Using struct name_entry also allows the use of tree_entry_len() for
finding the path length instead of strlen(), which may be slightly
more efficient.

Also unify the handling of missing entries in one of the two trees
(i.e. added or removed files): Just set cmp appropriately first, no
matter if we ran off the end of a tree or if we actually have two
entries to compare, and check its value a bit later without
duplicating the handler code.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0b9b3517fde28c2aa3c502012e8c796b71bdce53 [formerly 0b9b3517fde28c2aa3c502012e8c796b71bdce53 [formerly d8febde370aa464a5208cf094f306343e4ecb6dc]]
Former-commit-id: 13c54a96809eebdec1a389f5e286f91e00945c55
Former-commit-id: beacfb837a440aa0d81d78694f925a5272bebc05