git-web--browse: recognize any TERM_PROGRAM as a GUI terminal on OS X
It turns out that the presence of SECURITYSESSIONID is not sufficient
for detecting the presence of a GUI under Mac OS X.  SECURITYSESSIONID
appears to only be set when the user has Screen Sharing enabled.
Disabling Screen Sharing and relaunching the shell showed that the
variable was missing, at least under Mac OS X 10.6.8.

On the other hand, TERM_PROGRAM seems to be set for any terminals on
OS X, so just check it is set to something, instead of hardcoding
"Apple_Terminal" and missing other terminals such as iTerm.app.

Signed-off-by: John Szakmeister <john@szakmeister.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2ba2e806cf48ef4e7d7957b15b461a845b0e9f91 [formerly 2ba2e806cf48ef4e7d7957b15b461a845b0e9f91 [formerly be537e43ca8c6e38a3cf7024b28d6daf54611132]]
Former-commit-id: 5a2d9daad5700ac82ee6f674fe0d513681dba7bb
Former-commit-id: e16aee114dda983146b19f2b011f47f5567a681e