merge-one-file: force content conflict for "both sides added" case
Historically, we tried to be lenient to "both sides added, slightly
differently" case and as long as the files can be merged using a
made-up common ancestor cleanly, since f7d24bbefb06 (merge with
/dev/null as base, instead of punting O==empty case, 2005-11-07).

This was later further refined to use a better made-up common file
with fd66dbf5297a (merge-one-file: use empty- or common-base
condintionally in two-stage merge., 2005-11-10), but the spirit has
been the same.

But the original fix in f7d24bbefb06 to avoid punting on "both sides
added" case had a code to unconditionally error out the merge.  When
this triggers, even though the content-level merge can be done
cleanly, we end up not saying "content conflict" in the message, but
still issue the error message, showing "ERROR: in <pathname>".

Move that "always fail for add/add conflict" logic a bit higher to
fix this.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 505b8b13bbce26639800a330a161479a80f0d18e [formerly 505b8b13bbce26639800a330a161479a80f0d18e [formerly 4fa5c0591a7b5c71c8ccf71d8401c3337b7867ad]]
Former-commit-id: 8d037c064c4de9f1df6c11afda6f513c0f1c2ff5
Former-commit-id: 614bfd05319a95bff348a0b011c250e56b34dfc4