stream_blob_to_fd: detect errors reading from stream
We call read_istream, but never check its return value for
errors. This can lead to us looping infinitely, as we just
keep trying to write "-1" bytes (and we do not notice the
error, as we simply check that write_in_full reports the
same number of bytes we fed it, which of course is also -1).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1025ac79a9620e795cf24d1bc5c68b523db7c9ef [formerly 1025ac79a9620e795cf24d1bc5c68b523db7c9ef [formerly 45d4bdae5906cfe6b7cb1ba1cec82fd80381e07e]]
Former-commit-id: e5aa48e7215cf9bb014dc221306e6d4287c28fc9
Former-commit-id: 190354d43f18a58b6bb3d98af714706e632e7373