clone: die on errors from unpack_trees
When clone is populating the working tree, it ignores the
return status from unpack_trees; this means we may report a
successful clone, even when the checkout fails.

When checkout fails, we may want to leave the $GIT_DIR in
place, as it might be possible to recover the data through
further use of "git checkout" (e.g., if the checkout failed
due to a transient error, disk full, etc). However, we
already die on a number of other checkout-related errors, so
this patch follows that pattern.

In addition to marking a now-passing test, we need to adjust
t5710, which blindly assumed it could make bogus clones of
very deep alternates hierarchies. By using "--bare", we can
avoid it actually touching any objects.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4c52d6ea079e41456dbcb94fa06e33cdafa93987 [formerly 4c52d6ea079e41456dbcb94fa06e33cdafa93987 [formerly 0aac7bb287645dd72ad8ad6b805128b8ff7f111f]]
Former-commit-id: b29c8f3a93c19fd0e9527b41d59d32d10b172b70
Former-commit-id: 086d34902f6d05a30c6c2ac48d71b20a67a67b12