clone: run check_everything_connected
When we fetch from a remote, we do a revision walk to make
sure that what we received is connected to our existing
history. We do not do the same check for clone, which should
be able to check that we received an intact history graph.

The upside of this patch is that it will make clone more
resilient against propagating repository corruption. The
downside is that we will now traverse "rev-list --objects
--all" down to the roots, which may take some time (it is
especially noticeable for a "--local --bare" clone).

Note that we need to adjust t5710, which tries to make such
a bogus clone. Rather than checking after the fact that our
clone is bogus, we can simplify it to just make sure "git
clone" reports failure.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a34779c066fe7f68538156c4ccf5325f998e0a4a [formerly 502c94e08859f9d16de9dbce15e37c04acd3627b] [formerly 502c94e08859f9d16de9dbce15e37c04acd3627b [formerly 0433ad128c59f233046b3f8a68246ca3a8a77af8]]
Former-commit-id: 213ae4a60e66591a64d008ce0ebd22bd375dccb7
Former-commit-id: 44e7351daef20d8d66df9a62b313952173a1873c