streaming_write_entry: propagate streaming errors
When we are streaming an index blob to disk, we store the
error from stream_blob_to_fd in the "result" variable, and
then immediately overwrite that with the return value of
"close". That means we catch errors on close (e.g., problems
committing the file to disk), but miss anything which
happened before then.

We can fix this by using bitwise-OR to accumulate errors in
our result variable.

While we're here, we can also simplify the error handling
with an early return, which makes it easier to see under
which circumstances we need to clean up.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 531b9f5ef0f8307d1596911339ee916cd5d997a8 [formerly a875cd7057c310c58047e55f20d92d3c16b9565b] [formerly a875cd7057c310c58047e55f20d92d3c16b9565b [formerly d9c31e14d0aafdd45a382d01fcfd66c65a5f4b95]]
Former-commit-id: f561c621ba54e574fc25697d222c142f2dc0a47b
Former-commit-id: d14077ce57f44076b3baad182041b9a0c08802bf