clone: leave repo in place after checkout errors
If we manage to clone a remote repository but run into an
error in the checkout, it is probably sane to leave the repo
directory in place. That lets the user examine the situation
without spending time to re-clone from the remote (which may
be a lengthy process).

Rather than try to convert each die() from the checkout code
path into an error(), we simply set a flag that tells the
"remove_junk" atexit function to print a helpful message and
leave the repo in place.

Note that the test added in this patch actually passes
without the code change. The reason is that the cleanup code
is buggy; we chdir into the working tree for the checkout,
but still may use relative paths to remove the directories
(which means if you cloned into "foo", we would accidentally
remove "foo" from the working tree!).  There's no point in
fixing it now, since this patch means we will never try to
remove anything after the chdir, anyway.

[jc: replaced the message with a more succinct version from
Jonathan]

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1d9392b5e6d08966258bf431d2ad7710ab5e8285 [formerly 1d9392b5e6d08966258bf431d2ad7710ab5e8285 [formerly d3b34622f699ff14646de4ec1b1ab9afb0bcb056]]
Former-commit-id: fa5a499470ae04d1caff804b7f965a0ab5d22199
Former-commit-id: fb837d82ce9d4c0b4f4674c64b6aee11c6d0fbd3