gitk: Move hard-coded colors to .gitk
The Preferences dialog gives control of the colors of some elements of
the gitk user interface, but many are hard-coded in the gitk script.
In order to allow these to be customized through the gitk config
file, these other colors are stored in variables which can be set
in the config file, thus providing a way for color schemes to be stored
and shared.

For win32, this makes the default foreground color that of window text
rather than button text.

Signed-off-by: Gauthier Östervall <gauthier@ostervall.se>
[paulus@samba.org: Reworded commit message to be clearer,
 changed filesepfgcolor to black]
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 8b8e087850f892156659d6be99ef794ef5bf03b6 [formerly 8b8e087850f892156659d6be99ef794ef5bf03b6 [formerly 252c52df9d8d34692cb98a001491dd448e705901]]
Former-commit-id: 309afd5d3c17d0d80297122a1b0b5bfebb2fe1c3
Former-commit-id: 3325495cfa75b360af493763889a3ee1abec8692