Export rewrite_parents() for 'log -L'
The function rewrite_one is used to rewrite a single
parent of the current commit, and is used by rewrite_parents
to rewrite all the parents.

Decouple the dependence between them by making rewrite_one
a callback function that is passed to rewrite_parents. Then
export rewrite_parents for reuse by the line history browser.

We will use this function in line-log.c.

Signed-off-by: Bo Yang <struggleyb.nku@gmail.com>
Signed-off-by: Thomas Rast <trast@student.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: babff521ff1fc292974d95fea5e363cbe5064717 [formerly c5e72e9682ff2e08d5e1f17c7e57c8da8f597758] [formerly c5e72e9682ff2e08d5e1f17c7e57c8da8f597758 [formerly c7edcae06ed90ec0567179a78d7d1170e610c042]]
Former-commit-id: 7a4297f341d2427fda99e7a96fbe8755c5c29aa4
Former-commit-id: 0cdf86a594e81d8be6b5d76e562ad127dae8af3a