Speed up log -L... -M
So far log -L only used the implicit diff filtering by pathspec.  If
the user specifies -M, we cannot do that, and so we simply handed the
whole diff queue (which is approximately 'git show --raw') to
diffcore_std().

Unfortunately this is very slow.  We can optimize a lot if we throw
out files that we know cannot possibly be interesting, in the same
spirit that the pathspec filtering reduces the number of files.

However, in this case, we have to be more careful.  Because we want to
look out for renames, we need to keep all filepairs where something
was deleted.

This is a bit hacky and should really be replaced by equivalent
support in --follow, and just using that.  However, in the meantime it
speeds up 'log -M -L' by an order of magnitude.

Signed-off-by: Thomas Rast <trast@student.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 030afc32c916e14c462269c3ca5227cffa245651 [formerly b63ef643f8cf2d9b76db1f7729795c03296093d2] [formerly b63ef643f8cf2d9b76db1f7729795c03296093d2 [formerly 39410bf0bf81cb3f76fb9bfa8308cd672bc1442c]]
Former-commit-id: b5efdc31e876ce97388c4a3315b6df0f05141c84
Former-commit-id: eaee96aa5afcfa20b20bdfa39281ee7f66b3cf90