dir.c::match_pathname(): adjust patternlen when shifting pattern
If we receive a pattern that starts with "/", we shift it
forward to avoid looking at the "/" part. Since the prefix
and patternlen parameters are counts of what is in the
pattern, we must decrement them as we increment the pointer.

We remembered to handle prefix, but not patternlen. This
didn't cause any bugs, though, because the patternlen
parameter is not actually used. Since it will be used in
future patches, let's correct this oversight.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c88f5d990ada1a93a4df8da6b169295f987d362d [formerly e18a2ce7f72f12ec59c2189474c74c247aef8b42] [formerly e18a2ce7f72f12ec59c2189474c74c247aef8b42 [formerly 982ac87316a1cf5126888157bdcbfa32268ebe47]]
Former-commit-id: f4c8f991604037134f9d0837e22d5a016b1e5a19
Former-commit-id: 3c5af3e2ce6e5c43bc7b90b11ece1efbe618fd96