dir.c::match_pathname(): pay attention to the length of string parameters
This function takes two counted strings: a <pattern, patternlen> pair
and a <pathname, pathlen> pair. But we end up feeding the result to
fnmatch, which expects NUL-terminated strings.

We can fix this by calling the fnmatch_icase_mem function, which
handles re-allocating into a NUL-terminated string if necessary.

While we're at it, we can avoid even calling fnmatch in some cases. In
addition to patternlen, we get "prefix", the size of the pattern that
contains no wildcard characters. We do a straight match of the prefix
part first, and then use fnmatch to cover the rest. But if there are
no wildcards in the pattern at all, we do not even need to call
fnmatch; we would simply be comparing two empty strings.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c3bd5d6a233e08f9e48a7b9515e13f6b744c91ac [formerly b44ea9ab5debfe7a82e44bf8736ba7b235ad5679] [formerly b44ea9ab5debfe7a82e44bf8736ba7b235ad5679 [formerly ab3aebc15c13d083013591777688cecbcb703fb2]]
Former-commit-id: e2e8be59ea404e12a16b49ddf89477eb4381e80f
Former-commit-id: 2795a2d16222d5f4c6a4f1ee9a3d50869f91ad72