attr.c::path_matches(): special case paths that end with a slash
The function is given a string that ends with a slash to signal that
the path is a directory to make sure that a pattern that ends with a
slash (i.e. MUSTBEDIR) can tell directories and non-directories
apart.  However, the pattern itself (pat->pattern and
pat->patternlen) that came from such a MUSTBEDIR pattern is
represented as a string that ends with a slash, but patternlen does
not count that trailing slash. A MUSTBEDIR pattern "element/" is
represented as a counted string <"element/", 7> and this must match
match pathname "element/".

Because match_basename() and match_pathname() want to see pathname
"element" to match against the pattern <"element/", 7>, reduce the
length of the path to exclude the trailing slash when calling
these functions.

Signed-off-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: dfe26cdc1c25b6bc38bbea79456b98be349cdbc5 [formerly dfe26cdc1c25b6bc38bbea79456b98be349cdbc5 [formerly dc09e9ec43c09cdf803cc4f39f1fcb8ebcf80eb1]]
Former-commit-id: a843ad6dc8487907566a76907156e0e896b58654
Former-commit-id: 7cff4b9b1736d613b19414195de4a20578ca2666