t7800: run --dir-diff tests with and without symlinks
Currently the difftool --dir-diff tests may or may not use symlinks
depending on the operating system on which they are run.  In one case
this has caused a test failure to be noticed only on Windows when the
test also fails on Linux when difftool is invoked with --no-symlinks.

Rewrite these tests so that they do not depend on the environment but
run explicitly with both --symlinks and --no-symlinks, protecting the
--symlinks version with a SYMLINKS prerequisite.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5cf5daa3edf731796aa1f37617e6ead4e1655f25 [formerly 5cf5daa3edf731796aa1f37617e6ead4e1655f25 [formerly e01afdb74b395e63a6ee8feb8cb3a6ee470e2085]]
Former-commit-id: da00f67eb1087a09d4d87b978620852fb599918b
Former-commit-id: 14c71faf245cb053399a9eacb610665b6dc9e72c