difftool: don't overwrite modified files
After running the user's diff tool, git-difftool will copy any files
that differ between the working tree and the temporary tree.  This is
useful when the user edits the file in their diff tool but is wrong if
they edit the working tree file while examining the diff.

Instead of copying unconditionally when the files differ, create and
index from the working tree files and only copy the temporary file back
if it was modified and the working tree file was not.  If both files
have been modified, print a warning and exit with an error.

Note that we cannot use an existing index in git-difftool since those
contain the modified files that need to be checked out but here we are
looking at those files which are copied from the working tree and not
checked out.  These are precisely the files which are not in the
existing indices.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a36d667504c62f1547096c604ca22bc394b30493 [formerly a36d667504c62f1547096c604ca22bc394b30493 [formerly 67aa147af7c699348ba8c3afe9f7ee88aa043cd0]]
Former-commit-id: 3b1e98b7bf5f214ed9935ab7ad1cb7a96db334e2
Former-commit-id: 54f9cb161a1a4a6c39f9bcaf1397e2cc7c808889