rev-parse: clarify documentation for the --verify option
The old version could be read to mean that the argument has to refer
to a valid object, but that is incorrect:

* the object is not necessarily read (e.g., to check for corruption)

* if the argument is a 40-digit string of hex digits, then it is
  accepted whether or not is is the name of an existing object.

So reword the explanation to be less ambiguous.

Also fix the examples involving --verify: to be sure that the argument
refers to a commit (rather than some other kind of object), the
argument has to be suffixed with "^{commit}".  This trick is not
possible in the example involving --default, so don't imply that it is
exactly the same as the previous example.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 50b236533963f91e36c4b19a7c869bfd809e78f6 [formerly 50b236533963f91e36c4b19a7c869bfd809e78f6 [formerly 2db60670ba0c6dcaa42ed420f337851eea800bbc]]
Former-commit-id: 5ae645a06fea8894a505fbf1b5fc41d912052006
Former-commit-id: fa2c5254321a8f7fd17c83bc4bb3e56b31825416