git-svn: avoid self-referencing mergeinfo
When svn.pushmergeinfo is set, the target branch is included in the
mergeinfo if it was previously merged into one of the source branches.
SVN does not do this.

Remove merge target branch path from resulting mergeinfo when
svn.pushmergeinfo is set to better match the behavior of SVN. Update the
svn-mergeinfo-push test.

[ew: 80 columns]

Signed-off-by: Michael Contreras <michael@inetric.com>
Reported-by: Avishay Lavie <avishay.lavie@gmail.com>
Acked-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 86cfe8ad7a294a398f6572a6f6722acfe23539bc [formerly 86cfe8ad7a294a398f6572a6f6722acfe23539bc [formerly e234ac9d47b6618f7b7338f95f34d0b70782e4cb]]
Former-commit-id: b9c7eb70bfbae836d660b947cc00fc9aef0ddbd5
Former-commit-id: f8af3f41a3b1edcad900231dfafd364cc60f5440