tests --valgrind: provide a mode without --track-origins
With --valgrind=memcheck-fast, the tests run under memcheck but
without the autodetected --track-origins.  If you just run valgrind to
see *if* there is any memory issue with your program, the extra
information is not needed, and it comes at a roughly 30% hit in
runtime.

While it is possible to achieve the same through GIT_VALGRIND_OPTIONS,
this should be more discoverable and hopefully encourage more users to
run their tests with valgrind.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Acked-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a21f88cea76d0ee6005cde536093905cb3bdbb16 [formerly a21f88cea76d0ee6005cde536093905cb3bdbb16 [formerly 95d9d5ec75d3549f00cac29cd33be899cfb4693a]]
Former-commit-id: 98e2665f2929398a638f9cc9c3dbbdfd5980d771
Former-commit-id: d9ba768c30bfa30d1d4e67679f0f3dbca6773f7a