send-email: use "return;" not "return undef;" on error codepaths
All the callers of "ask", "extract_valid_address", and "validate_patch"
subroutines assign the return values from them to a single scalar:

	$var = subr(...);

and "return undef;" in these subroutine can safely be turned into a
simpler "return;".  Doing so will also future-proof a new caller that
mistakenly does this:

    @foo = ask(...);
    if (@foo) { ... we got an answer ... } else { ... we did not ... }

Note that we leave "return undef;" in validate_address on purpose,
even though Perlcritic may complain.  The primary "return" site of
the function returns whatever is in the scalar variable $address, so
it is pointless to change only the other "return undef;" to "return".
The caller must be prepared to see an array with a single undef as
the return value from this subroutine anyway.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1b9808b6070541a72f19b735e4ef57b53ced701f [formerly 1b9808b6070541a72f19b735e4ef57b53ced701f [formerly 622bc9309155b05af1a0d85bfb643bf6280eba35]]
Former-commit-id: ba5109520e83badf0b6b5d1ecc954db553882ff6
Former-commit-id: 4e8464b264b08b668846b4cc9164f1c2b6de03c9