send-email: use the three-arg form of open in recipients_cmd
Perlcritic does not want to see the trailing pipe in the two-args
form of open(), i.e.

	open my $fh, "$cmd \Q$file\E |";

If $cmd were a single-token command name, it would make a lot more
sense to use four-or-more-args form "open FILEHANDLE,MODE,CMD,ARGS"
to avoid shell from expanding metacharacters in $file, but we do
expect multi-word string in $to_cmd and $cc_cmd to be expanded by
the shell, so we cannot rewrite it to

	open my $fh, "-|", $cmd, $file;

for extra safety.  At least, by using this in the three-arg form:

	open my $fh, "-|", "$cmd \Q$file\E";

we can silence Perlcritic, even though we do not gain much safety by
doing so.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a0ff746bd2f7e4b0b080d9500e6ee722f6d884bf [formerly a0ff746bd2f7e4b0b080d9500e6ee722f6d884bf [formerly a47eab03f613fa55b9e690d5354e95bc165dceee]]
Former-commit-id: c1558e210b0d188b69efcf38c3b60696a6a9773d
Former-commit-id: d7f4d8405b7ac16ba55a93ae812dcb634691ac4f