t5516 (fetch-push): drop implicit arguments from helper functions
Many of the tests in t5516 look like:

  mk_empty &&
  git push testrepo ... &&
  check_push_result $commit heads/master

It's reasonably easy to see what is being tested, with the
exception that "testrepo" is a magic global name (it is
implicitly used in the helpers, but we have to name it
explicitly when calling git directly). Let's make it
explicit when call the helpers, too. This is slightly more
typing, but makes the test snippets read more naturally.

It also makes it easy for future tests to use an alternate
or multiple repositories, without a proliferation of helper
functions.

[rr: fixed sloppy quoting]

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Reviewed-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a69bcc9dda0b94612de53a5e1af8e4673e16f634 [formerly a69bcc9dda0b94612de53a5e1af8e4673e16f634 [formerly 2e433b7895798d2936d9708da9f99bd190cd39fc]]
Former-commit-id: 3ae410cdc0e0f76f02c0047234f137cfe6122b46
Former-commit-id: c8bedb246527d1ca4ded699a1dfcc1509b002a18