filter-branch: return to original dir after filtering
The first thing filter-branch does is to create a temporary
directory, either ".git-rewrite" in the current directory
(which may be the working tree or the repository if bare),
or in a directory specified by "-d". We then chdir to
$tempdir/t as our temporary working directory in which to run
tree filters.

After finishing the filter, we then attempt to go back to
the original directory with "cd ../..". This works in the
.git-rewrite case, but if "-d" is used, we end up in a
random directory. The only thing we do after this chdir is
to run git-read-tree, but that means that:

  1. The working directory is not updated to reflect the
     filtered history.

  2. We dump random files into "$tempdir/.." (e.g., if you
     use "-d /tmp/foo", we dump junk into /tmp).

Fix it by recording the full path to the original directory
and returning there explicitly.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bd85292732d98b1f75140a5767d0cc6f2776a9b2 [formerly 70415114d55de11c7a43b7ead4698f9e8e26c91b] [formerly 70415114d55de11c7a43b7ead4698f9e8e26c91b [formerly 97276019bb20829c97528b53dc453a37177c35bb]]
Former-commit-id: a77863be6b18794e5d2ef8ab94959a2dc7ad7f28
Former-commit-id: 01e401a80916dfe9cc4d6d5832541c2d4a0efe5d