t3200: test --set-upstream-to with bogus refs
These tests pass with the current code, but let's make sure
we don't accidentally break the behavior in the future.

Note that our tests expect failure when we try to set the
upstream to or from a missing branch. Technically we are
just munging config here, so we do not need the refs to
exist. But seeing that they do exist is a good check that
the user has not made a typo.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4a35e68393bafa88bf54cc99db97088a7f681717 [formerly 4a35e68393bafa88bf54cc99db97088a7f681717 [formerly 8a3e5ecdaa9fc45d7af58eb95bba0f39ca625a8d]]
Former-commit-id: cdaddcd3e409e6e8f4d92bb9b7b05f791fa13d07
Former-commit-id: 319532d57fbcfded7d63766af08a39ed4a2e7286