branch: improve error message for missing --set-upstream-to ref
If we are trying to set the upstream config for a branch,
the create_branch function will check both that the name
resolves as a ref, and that it is either a local or
remote-tracking branch.

However, before we do so we run get_sha1 on it to find out
whether it resolves at all (since the create_branch function
is also used to create actual branches, it wants to know
where to start the new branch). This means that if you feed
a ref that does not exist to "branch --set-upstream-to",
rather than getting a helpful message about tracking, you
only get "not a valid object name".

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ae00f7d5dbb7cbcfd0326772456654ccf5353021 [formerly ae00f7d5dbb7cbcfd0326772456654ccf5353021 [formerly a5e91c722cceb667ad05a13f9fde150cc1dbe9aa]]
Former-commit-id: c1f7070a314b75489dc74f9f005bd9cf2ee6dc4d
Former-commit-id: 18e0e80bef1ae91b59684eec7c8b39af278f3671