builtin/help.c: split "-a" processing into two
"help -a" (help all) gives the list of available commands and then
further gives hints on the use of "git help".   Separate these into
two steps, because we will add "help -g" (help guides) that want to
also show the overall hints after it is done.

While at it, change the definition of the "-a" option to use OPT_BOOL,
not the deprecated OPT_BOOLEAN.  We do not behave differently when
the user gives the "-a" option multiple times, e.g. "git help -a -a".

Signed-off-by: Philip Oakley <philipoakley@iee.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 925ebe375ad2883dc065b458969f0e86afa042b2 [formerly 925ebe375ad2883dc065b458969f0e86afa042b2 [formerly 15f7d4943807a50e01d2fd85bcdbd44361808dd2]]
Former-commit-id: f1b5533f90a54a3200146258d32b468ab023e3b5
Former-commit-id: 8603bb99537acaa5ed52c6c36199f70f66be6d46