diffcore-break: don't divide by zero
When the source file is empty, the calculation of the merge score
results in a division by zero.  In the situation:

     == preimage ==             == postimage ==

     F (empty file)             F (a large file)
                                E (a new empty file)

it does not make sense to consider F->E as a rename, so it is better not
to break the pre- and post-image of F.

Bail out early in this case to avoid hitting the divide-by-zero.  This
causes the merge score to be left at zero.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 65043e8cbcfce4b48c010600e00e07ac472cb5a9 [formerly 65043e8cbcfce4b48c010600e00e07ac472cb5a9 [formerly e7b00c57640fa2bf11e4218b588f51cc5a9a754e]]
Former-commit-id: aebee65aaa12d47be3827bd216d892bdb9007115
Former-commit-id: 122c35fb2f32f183dd9af1cb8eace3dfb50b5274