rerere forget: do not segfault if not all stages are present
The loop that fills in the buffers that are later passed to the merge
driver exits early when not all stages of a path are present in the index.
But since the buffer pointers are not initialized in advance, the
subsequent accesses are undefined.

Initialize buffer pointers in advance to avoid undefined behavior later.

That is not sufficient, though, to get correct operation of handle_cache().
The function replays a conflicted merge to extract the part inside the
conflict markers. As written, the loop exits early when a stage is missing.
Consequently, the buffers for later stages that would be present in the
index are not filled in and the merge is replayed with incomplete data.

Fix it by investigating all stages of the given path.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 293dc9ec9e55bc7d98eca4cef25545f35b6606ce [formerly 293dc9ec9e55bc7d98eca4cef25545f35b6606ce [formerly b9e31f59478eeae3e4230308f6fc06713c6fc547]]
Former-commit-id: dc0039075ecb68f1be83f9f224ad65674e9913bf
Former-commit-id: 04ed9b41fe79024f46b2a454201522c3c48fc6b2