rm: do not complain about d/f conflicts during deletion
If we used to have an index entry "d/f", but "d" has been
replaced by a non-directory entry, the user may still want
to run "git rm" to delete the stale index entry. They could
use "git rm --cached" to just touch the index, but "git rm"
should also work: we explicitly try to handle the case that
the file has already been removed from the working tree.

However, because unlinking "d/f" in this case will not yield
ENOENT, but rather ENOTDIR, we do not notice that the file
is already gone. Instead, we report it as an error.

The simple solution is to treat ENOTDIR in this case exactly
like ENOENT; all we want to know is whether the file is
already gone, and if a leading path is no longer a
directory, then by definition the sub-path is gone.

Reported-by: jpinheiro <7jpinheiro@gmail.com>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 138dff4cf5520a89028bafb5a07e8c95ff59f782 [formerly 138dff4cf5520a89028bafb5a07e8c95ff59f782 [formerly 9a6728d4d14a08791861a63b7567ee0b50e59f70]]
Former-commit-id: 87ec3a3ca1324c71df4d6de7d1d006518648086a
Former-commit-id: 803f37b1acb18a2e14cbdf3b1f8000dc813b3017