t3600: test behavior of reverse-d/f conflict
The previous commit taught "rm" that it is safe to consider
"d/f" removed when "d" has become a non-directory. This
patch adds a test for the opposite: a file "d" that becomes
a directory.

In this case, "git rm" does need to complain, because we
should not be removing arbitrary content under "d". Git
already behaves correctly, but let's make sure that remains
the case by protecting the behavior with a test.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d0dfbe321bfb1dbe6be70f6d2f2613e2869a8efe [formerly 3685d01228eca82541c87c0e02fc67e104737b9c] [formerly 3685d01228eca82541c87c0e02fc67e104737b9c [formerly 96ec8ee92aad52b6556e87d0166a7b9c4cd7c324]]
Former-commit-id: 066a6992a4956b0154abdd9bbf008af6de8d5aff
Former-commit-id: 2dcabf9529deed7d2ce72be110ced847b630672b