diffcore-pickaxe: remove unnecessary call to get_textconv()
The fill_one() function is responsible for finding and filling the
textconv filter as necessary, and is called by diff_grep() function
that implements "git log -G<pattern>".

The has_changes() function that implements "git log -S<block>" calls
get_textconv() for two sides being compared, before it checks to see
if it was asked to perform the pickaxe limiting.  Move the code
around to avoid this wastage.

After has_changes() calls get_textconv() to obtain textconv for both
sides, fill_one() is called to use them.

By adding get_textconv() to diff_grep() and relieving fill_one() of
responsibility to find the textconv filter, we can avoid calling
get_textconv() twice in has_changes().

With this change it's also no longer necessary for fill_one() to
modify the textconv argument, therefore pass a pointer instead of a
pointer to a pointer.

Signed-off-by: Simon Ruderich <simon@ruderich.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 81e44f4d231e1f029124709b980be9ab7abdccbf [formerly 36ce398056b1b7754d997067919b90050f5d1607] [formerly 36ce398056b1b7754d997067919b90050f5d1607 [formerly bc6158981b547db3d3eb754213467fed97109d57]]
Former-commit-id: 9ece4e0c8664d5732964f5e4a27b95c018974f7b
Former-commit-id: 011535cd6e84b70ea98e9db812e30d2d6dfa02b3