diffcore-pickaxe: remove fill_one()
fill_one is _almost_ identical to just calling fill_textconv; the
exception is that for the !DIFF_FILE_VALID case, fill_textconv gives us
an empty buffer rather than a NULL one. Since we currently use the NULL
pointer as a signal that the file is not present on one side of the
diff, we must now switch to using DIFF_FILE_VALID to make the same
check.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Simon Ruderich <simon@ruderich.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 441b8626a4e78ef3cedc55cc65b8739d7fe587d4 [formerly b715acf4ed450af1c8bd39026cc8d59c7e4c7b0c] [formerly b715acf4ed450af1c8bd39026cc8d59c7e4c7b0c [formerly 7cdb9b42c359000b1d3d604f847598afd015b7c7]]
Former-commit-id: 97a04d562b530f0af3e4290dae4f3f00c1542a41
Former-commit-id: 1fd458a3306fe3f6a38523737c8702cb41450f1b