diffcore-pickaxe: fix leaks in "log -S<block>" and "log -G<pattern>"
The diff_grep() and has_changes() functions had early return
codepaths for unmerged filepairs, which simply returned 0.  When we
taught textconv filter to them, one was ignored and continued to
return early without freeing the result filtered by textconv, and
the other had a failed attempt to fix, which allowed the planned
return value 0 to be overwritten by a bogus call to contains().

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c9d76763a4f92e9ff97ba1dfff40aa425da311af [formerly c9d76763a4f92e9ff97ba1dfff40aa425da311af [formerly 88ff684dd54f2a4793387f7162357005a4777ff2]]
Former-commit-id: d6fe0f50ad98f1021c1f43a64fb98039f4e33371
Former-commit-id: 1029121a13fc5d799f412ff3aca204280f8d04fa