diffcore-pickaxe: unify code for log -S/-G
The logic flow of has_changes() used for "log -S" and diff_grep()
used for "log -G" are essentially the same.  See if we have both
sides that could be different in any interesting way, slurp the
contents in core, possibly after applying textconv, inspect the
contents, clean-up and report the result.  The only difference
between the two is how "inspect" step works.

Unify this codeflow in a helper, pickaxe_match(), which takes a
callback function that implements the specific "inspect" step.

After removing the common scaffolding code from the existing
has_changes() and diff_grep(), they each becomes such a callback
function suitable for passing to pickaxe_match().

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d87d8e4ba9710944d1a427f50bce0cd246bb24e7 [formerly d87d8e4ba9710944d1a427f50bce0cd246bb24e7 [formerly 61690bf4a1ad499a673995b92cd8ab51104a431c]]
Former-commit-id: 97d12a7d84c9b1df5b927a86b5f0a229b34e8042
Former-commit-id: f5cc883a4438b3328c9a8fedc017ca046701893d