diff: allow unstuck arguments with --diff-algorithm
The argument to --diff-algorithm is mandatory, so there is no reason to
require the argument to be stuck to the option with '='.  Change this
for consistency with other Git commands.

Note that this does not change the handling of diff-algorithm in
merge-recursive.c since the primary interface to that is via the -X
option to 'git merge' where the unstuck form does not make sense.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7211b94e37c8df84ddcc2682a3c2217a92f87b1f [formerly 7211b94e37c8df84ddcc2682a3c2217a92f87b1f [formerly 0895c6d4c09045a3faeb7b652a844fc81c774631]]
Former-commit-id: 2826a3dae982f77009c01f2c87ea9df6c020e49d
Former-commit-id: e0c6219e9d222bc34bd9bf58c1536d094589e2c8