log -L: fix overlapping input ranges
The existing code was too defensive, and would trigger the assert in
range_set_append() if the user gave overlapping ranges.

The intent was always to define overlapping ranges as just the union
of all of them, as evidenced by the call to sort_and_merge_range_set().
(Which was already used, unlike what the comment said.)

Fix by splitting out the meat of range_set_append() to a new _unsafe()
function that lacks the paranoia.  sort_and_merge_range_set will fix
up the ranges, so we don't need the checks there.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e457463abbe74f7dff1a2388c8a8403184f2d26d [formerly e457463abbe74f7dff1a2388c8a8403184f2d26d [formerly 209618860c2627a4e134a15472587c574b328b40]]
Former-commit-id: d3578c5684836d3a1a39e59b5bb7529b027c7f67
Former-commit-id: cdc7f193de1c9a09d965440fcf27c283cce7a7aa