glossary: extend "detached HEAD" description
When we introduced the concept of "detached HEAD", we made sure that
commands that operate on the history of the current branch "just
work" in that state.  They update the HEAD to point at the new
history without affecting any branch when the HEAD is detached, just
like they update the tip of the "current branch" to point at the new
history when HEAD points at a specific branch.

As this is done as the natural extension for these commands, we did
not, we still do not, and we do not want to repeat "A detached HEAD
is updated without affecting any branch" when describing what each
and every one of these commands that operates "on the current branch"
does.

Add a blanket description to the glossary to cover them instead.
The general principle is that operations to update the branch work
on and affect the HEAD, while operations to update the information
about a branch do not.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9b8c2a282e58981b7ed74b2c34b9c2f84579acf1 [formerly 9b8c2a282e58981b7ed74b2c34b9c2f84579acf1 [formerly 1d77d249f9b0605ed21ca751c5dfb8550efcdfa3]]
Former-commit-id: 5d7e93157656b73b552adfe0fa21299e2c1be087
Former-commit-id: 461f501fc7b010db395053c55ffd1d5e08116bb8