show-branch: use strbuf instead of static buffer
When we generate relative names (e.g., "master~20^2"), we
format the name into a static buffer, then xstrdup the
result to attach it to the commit. Since the first thing we
add into the static buffer is the already-computed name of
the child commit, the names may get longer and longer as
the traversal gets deeper, and we may eventually overflow
the fixed-size buffer.

Fix this by converting the fixed-size buffer into a dynamic
strbuf.  The performance implications should be minimal, as
we end up allocating a heap copy of the name anyway (and now
we can just detach the heap copy from the strbuf).

Reported-by: Eric Roman <eroman@chromium.org>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aee51dab7e0821ac7a3b2b387f6321feaba74f97 [formerly aee51dab7e0821ac7a3b2b387f6321feaba74f97 [formerly aaa07e3eee6f42d0adeac826dfb52d5ade26defc]]
Former-commit-id: c2747e8655174a0d0ac0dcc85c41d0fe012768c2
Former-commit-id: 18a3a53c2ad40a70595c18ae131571a64c9a230d