http: add HTTP_KEEP_ERROR option
We currently set curl's FAILONERROR option, which means that
any http failures are reported as curl errors, and the
http body content from the server is thrown away.

This patch introduces a new option to http_get_strbuf which
specifies that the body content from a failed http response
should be placed in the destination strbuf, where it can be
accessed by the caller.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4121405538629b85533a6da3c24dc4338e6add2d [formerly 4121405538629b85533a6da3c24dc4338e6add2d [formerly 6d052d78d74e581dd93dd6328d3c214f469e34d7]]
Former-commit-id: 5260f2eaadd3673ea1f6fc3808acbe08a7c914b5
Former-commit-id: 2bc3a57bfcc2ce8b3dab911ca152e936b7428a97