remote-curl: consistently report repo url for http errors
When we report http errors in fetching the initial ref
advertisement, we show the full URL we attempted to use,
including "info/refs?service=git-upload-pack". While this
may be useful for debugging a broken server, it is
unnecessarily verbose and confusing for most cases, in which
the client user is not even the same person as the owner of
the repository.

Let's just show the repository URL; debugging can happen
with GIT_CURL_VERBOSE, which shows way more useful
information, anyway.

At the same time, let's also make sure to mention the
repository URL when we report failed authentication
(previously we said only "Authentication failed"). Knowing
the URL can help the user realize why authentication failed
(e.g., they meant to push to remote A, not remote B).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 3d0e073e79f14e567b4b34566216081437752cac [formerly 3d0e073e79f14e567b4b34566216081437752cac [formerly d5ccbe4dfb43d95d1d04252490fcd200e6aa5759]]
Former-commit-id: 8a15624568a5f8c744c3c8c9b5653d5365b49524
Former-commit-id: e74ccffba56b9b0b9ae94b68d43f45eb09d88042