http: re-word http error message
When we report an http error code, we say something like:

  error: The requested URL reported failure: 403 Forbidden while accessing http://example.com/repo.git

Everything between "error:" and "while" is written by curl,
and the resulting sentence is hard to read (especially
because there is no punctuation between curl's sentence and
the remainder of ours). Instead, let's re-order this to give
better flow:

  error: unable to access 'http://example.com/repo.git: The requested URL reported failure: 403 Forbidden

This is still annoyingly long, but at least reads more
clearly left to right.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d8ee8f84249f890d3598577a3be6a5157441a001 [formerly 6571dac6e065198e766f2aee156eec0b63f7c535] [formerly 6571dac6e065198e766f2aee156eec0b63f7c535 [formerly 39a570f26c953e8277a6a26c00da657b240dcdd4]]
Former-commit-id: 30a6df44df88af29b3c74d11702092766d4ba551
Former-commit-id: 01be2e360568c6e92b94ea144ce340d240fab8f8