remote-curl: die directly with http error messages
When we encounter an unknown http error (e.g., a 403), we
hand the error code to http_error, which then prints it with
error(). After that we die with the redundant message "HTTP
request failed".

Instead, let's just drop http_error entirely, which does
nothing but pass arguments to error(), and instead die
directly with a useful message.

So before:

  $ git clone https://example.com/repo.git
  Cloning into 'repo'...
  error: unable to access 'https://example.com/repo.git': The requested URL returned error: 403 Forbidden
  fatal: HTTP request failed

and after:

  $ git clone https://example.com/repo.git
  Cloning into 'repo'...
  fatal: unable to access 'https://example.com/repo.git': The requested URL returned error: 403 Forbidden

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d0692f55d97539d1fd48464539a6e89b6d090567 [formerly 355ed7c61818db6277224ca95b24e320021bc78c] [formerly 355ed7c61818db6277224ca95b24e320021bc78c [formerly de89f0b25a42238948787421c0253228c006f7fa]]
Former-commit-id: a41db60704f67649cd5b27391a96f614f5f97642
Former-commit-id: 10db3640ee33474c730da3565f4cf39f21aac6bb