gitremote-helpers(1): clarify refspec behaviour
The documentation says that "If no 'refspec' capability is advertised,
there is an implied `refspec *:*`" but this is only the case for the
"import" command.

Since there is a comment in transport-helper.c indicating that this
default is for historical reasons, change the documentation to clarify
that a refspec should always be specified.

Signed-off-by: John Keeping <john@keeping.me.uk>
Acked-by: Sverre Rabbelier <srabbelier@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: aaa228eeefe5bbcc9e0d9c7c731985d7eefe193b [formerly aaa228eeefe5bbcc9e0d9c7c731985d7eefe193b [formerly 9e7673ed7ff656738fdc02d1c529d363aaa98002]]
Former-commit-id: 4c1fbe9364b9d8c756359b98f6dfaeeccedb27d9
Former-commit-id: 8d482588426fa14c87d6fcf9261cf181ccdcd427