remote-helpers/test-bzr.sh: do not use "grep '\s'"
Using grep "devel\s\+3:" to find at least one whitspace is not
portable on all grep versions; not all grep versions understand "\s"
as a "whitespace".

Use a literal TAB followed by SPACE.

The + as a qualifier for "one or more" is not a basic regular
expression; use egrep instead of grep.

Signed-off-by: Torsten Bögershausen <tboegi@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 36ac6853fae73a516536d0e4c9e29d9b0bd659a9 [formerly 36ac6853fae73a516536d0e4c9e29d9b0bd659a9 [formerly 0e9b327227fb8f7c1e3a5bbaef19e25b4dfd6764]]
Former-commit-id: 435c69e2961a291115a07831da13f20c2fa5880e
Former-commit-id: 3c3de1505f0972c8a8bc1ae4c1d7eb344de6c3df