bundle: Accept prerequisites without commit messages
While explicitly stating that the commit message in a prerequisite
line is optional, we required all lines with 40 or more characters
to contain a space after the object name, bailing out if a line
consisted of an object name only. This was to allow bundling a
history to a commit without an message, but the code forgot that it
already called rtrim() to remove that whitespace.

As a workaround, only check for SP when the line has more than 40
characters.

Signed-off-by: Lukas Fleischer <git@cryptocrack.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5e466cde3b1481ecea9f4d436f4de745f138cf00 [formerly f9d6c54bac57fe9f5fbcc0890d83b8ffe024ae5e] [formerly f9d6c54bac57fe9f5fbcc0890d83b8ffe024ae5e [formerly 5446e33f35d6b27b6b760a59a6ded33fbd555190]]
Former-commit-id: 0669e5d1d172efb3b8833eb37e9a816c7186a0a6
Former-commit-id: 5875529af877e7f225391cf3605d174f2a644948