format-patch: improve head calculation for cover-letter
If we do it after the revision traversal we can be sure that this is
indeed a commit that will be processed (i.e. not a merge) and it's the
top most one (thus removing the NEEDSWORK comment, at least we show the
same as 'git diff --stat' output that appears in the cover-letter).

While we are at it, since we know there's nothing to generate, exit
sooner in all cases, like --cover-letter currently does.

Also, if there's nothing to generate and cover-letter is specified, a
different code-path might be triggered that is not currently covered in
the test-case, so add a test for it.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5e9f76a3940a25d8a8752f225670307f78fdd623 [formerly 7032c2378644bdbce28e237fc258408871326bb8] [formerly 7032c2378644bdbce28e237fc258408871326bb8 [formerly 80d35ca0aaeca654562974be6ea528e22bc76e8c]]
Former-commit-id: d4c4bf9bc4fd9c9d51d74931e0b085c275639fe4
Former-commit-id: ee6851ea2f36059a58903ac9a06a30f5c363dbe1