clone: Fix error message for reference repository
Do not report that an argument to clone's --reference option is not a
local directory.  Nothing checks for the existence or type of the path
as supplied by the user; checks are only done for particular contents of
the supposed directory, so we have no way to know the status of the
supplied path.  Telling the user that a directory doesn't exist when
that isn't actually known may lead him or her on the wrong path to
finding the problem.

Instead just state that the entered path is not a local repository which
is really all that is known about it.  It could be more helpful to state
the actual paths which were checked, but I believe that giving a good
description of that would be too verbose for a simple error message and
would be too dependent on implementation details.

Signed-off-by: Aaron Schrab <aaron@schrab.com>
Reviewed-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: f5ef2d7fb99bd8e78a9415be03cc49888873d80d [formerly f5ef2d7fb99bd8e78a9415be03cc49888873d80d [formerly 0658569eb03e743ef46abd06abd9d0d81d53e72c]]
Former-commit-id: 879104b86114ca643dfbbd662101f75e12ce27be
Former-commit-id: e4fb0260bf711205ac7203f746f2cbe11a4aa7d0