commit: allow associating auxiliary info on-demand
The "indegree" field in the commit object is only used while sorting
a list of commits in topological order, and wasting memory otherwise.

We would prefer to shrink the size of individual commit objects,
which we may have to hold thousands of in-core. We could eject
"indegree" field out from the commit object and represent it as a
dynamic table based on the decoration infrastructure, but the
decoration is meant for sparse annotation and is not a good match.

Instead, let's try a different approach.

 - Assign an integer (commit->index) to each commit we keep in-core
   (reuse the space of "indegree" field for it);

 - When running the topological sort, allocate an array of integers
   in bulk (called "slab"), use the commit->index as an index into
   this array, and store the "indegree" information there.

This does _not_ reduce the memory footprint of a commit object, but
the commit->index can be used as the index to dynamically associate
commits with other kinds of information as needed.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a3a0d108178b9392066a58b079ef7692743df6ec [formerly e466ade67b9f03c21da0ad59642d89153950aa73] [formerly e466ade67b9f03c21da0ad59642d89153950aa73 [formerly 96c4f4a370591b4796628abe18f27f0133b21954]]
Former-commit-id: f4ad137dd7a0f95e03877f6fd6a93d14caeff66c
Former-commit-id: ec7edfce5d3e9e146452b24f410569634c8c343c