completion: inline __gitcomp_1 to its sole callsite
There is no point in calling a separate function that is only used
in one place. Especially considering that there's no need to call
compgen, and we traverse the words ourselves both in __gitcompadd,
and __gitcomp_1.

Let's squash the functions together, and traverse only once.

This improves performance. For N number of words:

  == 1 ==
  original: 0.002s
  new: 0.000s
  == 10 ==
  original: 0.005s
  new: 0.001s
  == 100 ==
  original: 0.009s
  new: 0.006s
  == 1000 ==
  original: 0.027s
  new: 0.019s
  == 10000 ==
  original: 0.163s
  new: 0.151s
  == 100000 ==
  original: 1.555s
  new: 1.497s

No functional changes.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 630fcc66a93bf98284dab60f0cfd97f0ed80baf4 [formerly 630fcc66a93bf98284dab60f0cfd97f0ed80baf4 [formerly b4cfbc969cd910992a08b12dc612d19323668e09]]
Former-commit-id: a93dacf5ed3aba45a2955a2991c7858b1d43d72a
Former-commit-id: 4fea8b65a6eb5e0a59cc3b2812ceade89aa550cd