completion: small optimization
No need to calculate a new $c with a space if we are not going to do
anything it with it.

There should be no functional changes, except that a word "foo " with no
suffixes can't be matched. But $cur cannot have a space at the end
anyway. So it's safe.

Based on the code from SZEDER Gábor.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 14baac3bce34d6fa8864e1f6a1ef131ef2be2f16 [formerly 14baac3bce34d6fa8864e1f6a1ef131ef2be2f16 [formerly ddc996d767025a4ffcb52a813d32a5095f357e44]]
Former-commit-id: 7f5fa9be845ca8059984534baa8022d8881d7f7a
Former-commit-id: 925c6ce053be9ae39c1b67e1575f7b97b84f7d24