t5004: fix issue with empty archive test and bsdtar
bsdtar, which is the default tar on Mac OS X, handles empty archives
just fine but reports archives containing only a pax extended header
comment as damaged.  Work around the issue by explicitly generating
the archive for the tree and not the commit, which causes git archive
to omit the commit hash comment record from the tar file.

Reported-by: BJ Hargrave <bj@bjhargrave.com>
Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5d8ed42bc063e1f59fca591cff8ec7a2a9f7daf6 [formerly 5d8ed42bc063e1f59fca591cff8ec7a2a9f7daf6 [formerly 24676f02ba8bcf0f2c5cccd02825e15ca93bb5f4]]
Former-commit-id: 844544981d8eb614963b5753a1c47f86480e353e
Former-commit-id: 6bbba1a33deb9470b601e59dd68bed7276979910