strbuf: create strbuf_humanise_bytes() to show byte sizes
Humanization of downloaded size is done in the same function as text
formatting in 'process.c'. The code cannot be reused easily elsewhere.

Separate text formatting from size simplification and make the
function public in strbuf so that it can easily be used by other
callers.

We now can use strbuf_humanise_bytes() for both downloaded size and
download speed calculation. One of the drawbacks is that speed will
now look like this when download is stalled: "0 bytes/s" instead of
"0 KiB/s".

Signed-off-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 78d52131aa1fc9e8ea3c51622da01f5fd121c861 [formerly 78d52131aa1fc9e8ea3c51622da01f5fd121c861 [formerly 079b546a2940a84989893254a318c8414e13a13e]]
Former-commit-id: 9df6bdea4c4a290104235eb38f57514ea42b6f05
Former-commit-id: 6546fba2434c9b4ebcc90d43b90129cf0c25aaed