transport-helper: mention helper name when it dies
When we try to read from a remote-helper and get EOF or an
error, we print a message indicating that the helper died.
However, users may not know that a remote helper was in use
(e.g., when using git-over-http), or even what a remote
helper is.

Let's print the name of the helper (e.g., "git-remote-https");
this makes it more obvious what the program is for, and
provides a useful token for reporting bugs or searching for
more information (e.g., in manpages).

Signed-off-by: Jeff King <peff@peff.net>
Acked-by: Sverre Rabbelier <srabbelier@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fd40f4dbf0b8a93f4381e8340eb3ff3dd206f958 [formerly fd40f4dbf0b8a93f4381e8340eb3ff3dd206f958 [formerly c096955c5bb8e20186cc7c07d4d12b77ddcd01c6]]
Former-commit-id: db141edf7b04eacc0fcc6a8b47b565964da43356
Former-commit-id: 5398149802435bd4cfeabba5faa954d78c2d7190