check-ignore: add -n / --non-matching option
If `-n` or `--non-matching` are specified, non-matching pathnames will
also be output, in which case all fields in each output record except
for <pathname> will be empty.  This can be useful when running
check-ignore as a background process, so that files can be
incrementally streamed to STDIN, and for each of these files, STDOUT
will indicate whether that file matched a pattern or not.  (Without
this option, it would be impossible to tell whether the absence of
output for a given file meant that it didn't match any pattern, or
that the result simply hadn't been flushed to STDOUT yet.)

Signed-off-by: Adam Spiers <git@adamspiers.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: affdd53ad54eaf6d92dd6d2e3876d24989070562 [formerly affdd53ad54eaf6d92dd6d2e3876d24989070562 [formerly ae3caf4c912429c8e9755f2e10d2f0a7f6279c36]]
Former-commit-id: 11df60d9e473bc569833327d0adf496ef1ae3106
Former-commit-id: 612a9c21fae327145c63d7b36c9ad83764f6dc60