check-ignore: allow incremental streaming of queries via --stdin
Some callers, such as the git-annex web assistant, find it useful to
invoke git check-ignore as a persistent background process, which can
then have queries fed to its STDIN at any point, and the corresponding
response consumed from its STDOUT.  For this we need to invoke
check_ignore() once per line of standard input, and flush standard
output after each result.

The above use case suggests that empty STDIN is actually a reasonable
scenario (e.g. when the caller doesn't know in advance whether any
queries need to be fed to the background process until after it's
already started), so we make the minor behavioural change that "no
pathspec given." is no longer emitted in when STDIN is empty.

Even though check_ignore() could now be changed to operate on a single
pathspec, we keep it operating on an array of pathspecs since that is
a more convenient way of consuming the existing pathspec API.

Signed-off-by: Adam Spiers <git@adamspiers.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b1595b06403cf12f614b3b207dbfbe7688f924c2 [formerly 37f1b72afec1783c6e5d0120726090534a90daf7] [formerly 37f1b72afec1783c6e5d0120726090534a90daf7 [formerly 0c8e8c080b08632d6dbee33d093b5df648196b49]]
Former-commit-id: 747555ad2a71a9f4e1cd008d06e6ef94c8f2373c
Former-commit-id: 1ab2d5b823218730a6f25ed0e4b742ec499076db