Documentation: add caveats about I/O buffering for check-{attr,ignore}
check-attr and check-ignore have the potential to deadlock callers
which do not read back the output in real-time.  For example, if a
caller writes N paths out and then reads N lines back in, it risks
becoming blocked on write() to check-*, and check-* is blocked on
write back to the caller.  Somebody has to buffer; the pipe buffers
provide some leeway, but they are limited.

Thanks to Peff for pointing this out:

    http://article.gmane.org/gmane.comp.version-control.git/220534

Signed-off-by: Adam Spiers <git@adamspiers.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 72bc2aa9228da5d5439264f6e3dc4014e66d83c6 [formerly 72bc2aa9228da5d5439264f6e3dc4014e66d83c6 [formerly f1ed7fea7974218db15155f3b8d2e29d3fe0971c]]
Former-commit-id: 46d32756db103f99b23501340ceb61cedfd9477a
Former-commit-id: 59d93215df612af540d04264068fc33114fe1c7b