remote-hg: force remote push
Ideally we shouldn't do this, as it's not recommended in mercurial
documentation, but there's no other way to push multiple bookmarks (on
the same branch), which would be the behavior most similar to git.

At the same time, add a configuration option for the people that don't
want to risk creating new remote heads.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a3f91e2a49eda0fd398b695ab71b306ce28c0e31 [formerly a3f91e2a49eda0fd398b695ab71b306ce28c0e31 [formerly b0c3db860cac2215f07deeb9e91cbd2597fe41d0]]
Former-commit-id: d69584bebf937bc798a0189f56dcbf33ef89fffd
Former-commit-id: c07ed02ecf44f77fb1e6034baaa007c9464ba2e2