remote-hg: activate graphlog extension for hg_log()
The hg_log() test helper uses the "--graph" parameter that is
implemented by the GraphLog extension. If the extension is not activated
by the user, the parameter is not available. Activate the extension in
setup().

Also changes the way we grep the output in hg_log(). The pipe operator
can hide the return code of hg command. As a matter of fact, if log
fails because it doesn't know about "--graph", it doesn't report any
failure and let's you think everything worked.

Signed-off-by: Antoine Pelisse <apelisse@gmail.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c7935ca6c06ad99cb620359ae0c894878014dc66 [formerly c7935ca6c06ad99cb620359ae0c894878014dc66 [formerly 9a57988b3fac79e0458515a555392d00b759cf4c]]
Former-commit-id: 6119b61f9f107cf6b9b0db5a887fe2b26c3d3137
Former-commit-id: a04178c38a183d855bfc85b730d0f9206e8a633e