gitweb/INSTALL: Simplify description of GITWEB_CONFIG_SYSTEM
The flow of the text describing GITWEB_CONFIG_SYSTEM and
GITWEB_CONFIG_COMMON in gitweb/INSTALL is awkward.  "This is bad. Oh
the other hand, better is broken. Therefore we do this." forces
readers to make multiple guesses while reading: "ok, bad, so you
plan to change it and warn us about upcoming change?  oh, not that,
changing it is bad, so we have to live with it?  oh, not that, there
is another one that is common and that is what we can use".

Better rewrite said paragraph to avoid such a mental roller-coaster in
the first place.

Signed-off-by: Junio Hamano <gitster@pobox.com>
Signed-off-by: Jakub Narebski <jnareb@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a697a083abe01ed33cf5b7ae186059f46d28344d [formerly a697a083abe01ed33cf5b7ae186059f46d28344d [formerly 0d2f7d1c5ef20a6872ce1eacb054c576bb048057]]
Former-commit-id: a337b11bbd4356ab665da564a3b18b9fb20d96d6
Former-commit-id: 4799db339c260fd560ddeb8ead2003870b7d23e5