checkout: add --ignore-skip-worktree-bits in sparse checkout mode
"git checkout -- <paths>" is usually used to restore all modified
files in <paths>. In sparse checkout mode, this command is overloaded
with another meaning: to add back all files in <paths> that are
excluded by sparse patterns.

As the former makes more sense for day-to-day use. Switch it to the
default and the latter enabled with --ignore-skip-worktree-bits.

While at there, add info/sparse-checkout to gitrepository-layout.txt

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 47aee39f4b1bd805f983af977a842bbdb68ffecd [formerly 47aee39f4b1bd805f983af977a842bbdb68ffecd [formerly 08d595dc1cdf6f0d8e6022a69c4fcdd2fba628cf]]
Former-commit-id: f93d469ddc2c08d8d570a478a43dd5f46cdaecd2
Former-commit-id: 3b8913844ffe9925bb7af7eaa1fb3493c49a88a7