doc/http-backend: match query-string in apache half-auth example
When setting up a "half-auth" repository in which reads can
be done anonymously but writes require authentication, it is
best if the server can require authentication for both the
ref advertisement and the actual receive-pack POSTs. This
alleviates the need for the admin to set http.receivepack in
the repositories, and means that the client is challenged
for credentials immediately, instead of partway through the
push process (and git clients older than v1.7.11.7 had
trouble handling these challenges).

Since detecting a push during the ref advertisement requires
matching the query string, and this is non-trivial to do in
Apache, we have traditionally punted and instructed users to
just protect "/git-receive-pack$".  This patch provides the
mod_rewrite recipe to actually match the ref advertisement,
which is preferred.

While we're at it, let's add the recipe to our test scripts
so that we can be sure that it works, and doesn't get broken
(either by our changes or by changes in Apache).

Signed-off-by: Jeff King <peff@peff.net>
Acked-by: Jakub Narębski <jnareb@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2ebcdc7e35219e7f78d86b4a48e1468c0aa18f04 [formerly e3d389304cbf4770aae510c367fe6233c4e27b98] [formerly e3d389304cbf4770aae510c367fe6233c4e27b98 [formerly b0808819e5806f5ff01ffcc34db2796d180ad0d9]]
Former-commit-id: 6e9ff32c8e6d17af2ae2b7bf41fedd4f225def56
Former-commit-id: a648ae1e2d68fd2f526348567e59bb1b6a8e808a