commit-slab: avoid large realloc
Instead of using a single "slab" and keep reallocating it as we find
that we need to deal with commits with larger values of commit->index,
make a "slab" an array of many "slab_piece"s. Each access may need
two levels of indirections, but we only need to reallocate the first
level array of pointers when we have to grow the table this way.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c0fa60ea36c458e5371eec85b57969984a51be22 [formerly c0fa60ea36c458e5371eec85b57969984a51be22 [formerly 66eb375d3d334efa3f467775a5f2a647c131c4b1]]
Former-commit-id: 53aa91918e988276e65ac64f68f55fc9d0fc367a
Former-commit-id: 726157fe09999ce76e966355a38f088ea79a5490