transport-helper: pass --signed-tags=warn-strip to fast-export
Currently, attempting to push a signed tag to a remote helper which uses
fast-export results in the remote helper failing because the default
fast-export action for signed tags is "abort".  This is not helpful for
users because there is no way to pass additional arguments to
fast-export here, either from the remote helper or from the command
line.

In general, the signature will be invalidated by whatever transformation
a remote helper performs on a tag to push it to a repository in a
different format so the correct behaviour is to strip the tag.  Doing
this silently may surprise people, so use "warn-strip" to issue a
warning when a signed tag is encountered.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 48b478a858a3fe0c7d5b80fa60bb3810650bb72a [formerly 48b478a858a3fe0c7d5b80fa60bb3810650bb72a [formerly b8bd826f3a4196bc80a3c957503b421709b401ab]]
Former-commit-id: c091b09edcd96bfb95ad2b7f22aa98c0455eaa31
Former-commit-id: 526701ec42f615a3969f244532c88d153f46679a