t/test-lib.sh: fix TRASH_DIRECTORY handling
After the location of $TRASH_DIRECTORY is adjusted by
$TEST_OUTPUT_DIRECTORY, we go on to use the $test variable to make the
trash directory and cd into it.  This means that when
$TEST_OUTPUT_DIRECTORY is not "." and an absolute --root has not been
specified, we do not remove the trash directory once the tests are
complete (remove_trash is set to $TRASH_DIRECTORY).

Fix this by always referring to the trash directory as $TRASH_DIRECTORY.

Signed-off-by: John Keeping <john@keeping.me.uk>
Acked-by: Jeff King <peff@peff.net>
Acked-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c33ea84849ae26c6c5d1672486f51d771e4b51fc [formerly c33ea84849ae26c6c5d1672486f51d771e4b51fc [formerly 38b074de80759c95fec20b0dbc30614caccdfd24]]
Former-commit-id: deaa9cd1c56945fce14799ed2e5e030816992cd2
Former-commit-id: 63faf443039244162d9e6eab417211a2d2a1429f