t/test-lib.sh: drop "$test" variable
The $test variable is used as an interim buffer for
constructing $TRASH_DIRECTORY, and is almost compatible with
it (the exception being that $test has not been converted to
an absolute path). Let's get rid of it entirely so that
later code does not accidentally use it, thinking the two
are interchangeable.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 265522aab04f3c2c510ffa7cf081e8c233c0105e [formerly abf8bbc34027644967ca9e8eaf1bd34829bd868c] [formerly abf8bbc34027644967ca9e8eaf1bd34829bd868c [formerly 002d4ce8aa40644698721d59691d38f1df702f4b]]
Former-commit-id: b151facc152eca2921708553fa1d51c98b489779
Former-commit-id: 6bb0f876dc7312f5609a76438ec68b3c7dcf858f