glossary: improve description of SHA-1 related topics
The name of the hash function is "SHA-1", not "SHA1".

Also to people who look up "object name" in the glossary,
the details of which hash function is applied on what to
compute "object name" is not important but the fact that the
name is meant to be an unique identifier for the contents
stored in the object is.

Signed-off-by: Thomas Ackermann <th.acker@arcor.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0793d77683816cbae92a975b230639f359845755 [formerly 0793d77683816cbae92a975b230639f359845755 [formerly 3ab501209b46e83a377ac5a781a8f2fef7f4f30c]]
Former-commit-id: a5cc42c17fc1e825c0bb67c7e0c10446ef37b91f
Former-commit-id: 8ed0f01ac7c0839ef024afb0afaa6334c8f2e8df