glossary: improve definitions of refspec and pathspec
The exact definition of "refspec" can be found in git-fetch and
git-push manpages. So don't duplicate this here in the glossary.

Actually the definition of "pathspec" should be moved to a separate
file akin to the way it's done with "refspec". But this will only be
wortwhile when there's more to say about it. So for the time being
just improve the first sentence a little bit; fix the indentation of
the first paragraph after the bullet list and remove the one-item
list of magic signatures with its - for the user - unnecessary
introduction of "magic word 'top'".

Signed-off-by: Thomas Ackermann <th.acker@arcor.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5a6880c5399ebed13a6b2d0de4344ab0270dcf95 [formerly 063a7c8228f4c4f297da98632b9100625fa5de5f] [formerly 063a7c8228f4c4f297da98632b9100625fa5de5f [formerly 57148ebb30fd7ed72acf7b3fc6f59d7cf2d94643]]
Former-commit-id: cd34e42e8a816bc653b1867d18dda523317668a2
Former-commit-id: 7a85ee659bcb1f66f3fdcbee1ceb74d498ce958f