usage: allow pluggable die-recursion checks
When any git code calls die or die_errno, we use a counter
to detect recursion into the die functions from any of the
helper functions. However, such a simple counter is not good
enough for threaded programs, which may call die from a
sub-thread, killing only the sub-thread (but incrementing
the counter for everyone).

Rather than try to deal with threads ourselves here, let's
just allow callers to plug in their own recursion-detection
function. This is similar to how we handle the die routine
(the caller plugs in a die routine which may kill only the
sub-thread).

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 2c73e864c78bd14c4d41c97c70a53d103e614b32 [formerly 2c73e864c78bd14c4d41c97c70a53d103e614b32 [formerly c19a490e37181de8fa94ae1074af4b9f9a518f95]]
Former-commit-id: bc82d36652c5abd7dd36b7daa8124bec39b2274b
Former-commit-id: 697c50c0249abe5dce382c4e1d0a27d109f8eeb6