blame: handle broken commit headers gracefully
split_ident_line() can leave us with the pointers date_begin, date_end,
tz_begin and tz_end all set to NULL.  Check them before use and supply
the same fallback values as in the case of a negative return code from
split_ident_line().

The "(unknown)" is not actually shown in the output, though, because it
will be converted to a number (zero) eventually.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ba726d2cc5ddce88eeed9ead58456b03aaf3a802 [formerly 8fac8a3a8bd3866d106a71a91923554fb25301d8] [formerly 8fac8a3a8bd3866d106a71a91923554fb25301d8 [formerly de5abe9fe91a496d019d62abefe23df9d72fad30]]
Former-commit-id: 602da055e2de6edee91e531c41ad9122e803ead0
Former-commit-id: aa569ced5736bdc8d2572bc57095de61686c3f9b