date.c: add parse_expiry_date()
"git reflog --expire=all" tries to expire reflog entries up to the
current second, because the approxidate() parser gives the current
timestamp for anything it does not understand (and it does not know
what time "all" means).  When the user tells us to expire "all" (or
set the expiration time to "now"), the user wants to remove all the
reflog entries (no reflog entry should record future time).

Just set it to ULONG_MAX and to let everything that is older that
timestamp expire.

While at it, allow "now" to be treated the same way for callers that
parse expiry date timestamp with this function.  Also use an error
reporting version of approxidate() to report misspelled date.  When
the user says e.g. "--expire=mnoday" to delete entries two days or
older on Wednesday, we wouldn't want the "unknown, default to now"
logic to kick in.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 394467183ba100ec0665e5b0310ad9b365c8f1da [formerly 394467183ba100ec0665e5b0310ad9b365c8f1da [formerly 3d27b9b005f26b107227fc18b6648df329daee97]]
Former-commit-id: d8b9a0991e4396f3754edbb6c5f071c98a132808
Former-commit-id: 40d12b7ac9fc401f7a9e39256e01ac47415ad56d