transport-helper: warn when refspec is not used
For the modes that need it. In the future we should probably error out,
instead of providing half-assed support.

The reason we want to do this is because if it's not present, the remote
helper might be updating refs/heads/*, or refs/remotes/origin/*,
directly, and in the process fetch will get confused trying to update
refs that are already updated, or older than what they should be. We
shouldn't be messing with the rest of git.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4c80f263e9457e039a668317a32408e54a89b659 [formerly 4c80f263e9457e039a668317a32408e54a89b659 [formerly a93b4a09acfed0e2a006770d0196c74968e65c25]]
Former-commit-id: e55951e4d36ab449eff0b63f1d89a658bb81964e
Former-commit-id: 2afedd5bc212b81be6fb8efac039490e3415fc8b