pretty: two phase conversion for non utf-8 commits
Always assume format_commit_item() takes an utf-8 string for string
handling simplicity (we can handle utf-8 strings, but can't with other
encodings).

If commit message is in non-utf8, or output encoding is not, then the
commit is first converted to utf-8, processed, then output converted
to output encoding. This of course only works with encodings that are
compatible with Unicode.

This also fixes the iso8859-1 test in t6006. It's supposed to create
an iso8859-1 commit, but the commit content in t6006 is in UTF-8.
t6006 is now converted back in UTF-8 (the downside is we can't put
utf-8 strings there anymore).

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ed92ece625d16bc80dd63403af57a718d8e4cfb1 [formerly ed92ece625d16bc80dd63403af57a718d8e4cfb1 [formerly 7e77df39bf20f04b266f394a64c2fca54c1e5db0]]
Former-commit-id: 19a98ed4ee12f09f36a90f9f44a6b83dcbfa21ac
Former-commit-id: e9cc9eae86f186652cc39c5109ba0335595dc71b