pretty: support padding placeholders, %< %> and %><
Either %<, %> or %>< standing before a placeholder specifies how many
columns (at least as the placeholder can exceed it) it takes. Each
differs on how spaces are padded:

  %< pads on the right (aka left alignment)
  %> pads on the left (aka right alignment)
  %>< pads both ways equally (aka centered)

The (<N>) follows them, e.g. `%<(100)', to specify the number of
columns the next placeholder takes.

However, if '|' stands before (<N>), e.g. `%>|(100)', then the number
of columns is calculated so that it reaches the Nth column on screen.

Signed-off-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ade1b60b83b25eb094fdbf6b67fea563c4518690 [formerly ade1b60b83b25eb094fdbf6b67fea563c4518690 [formerly a57523428b4c08795e793356e12afc3de812f84a]]
Former-commit-id: 23fc00cbf90a59c4568c53873f6a46367c5b5f1a
Former-commit-id: 3d45dbbeeaa5ca3be4dde4e794b1d6cc5dbac439