t3200.39: tracking setup should fail if there is no matching refspec.
We are formalizing a requirement that any remote-tracking branch to be used
as an upstream (i.e. as an argument to --track), _must_ "belong" to a
configured remote by being matched by the "dst" side of a fetch refspec.

This patch encodes the new expected behavior of this test, and marks the
test with "test_expect_failure" in anticipation of a following patch to
introduce the new behavior.

Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d9377b4647a43bd8d5c5bf1b3fb14b6d071dd135 [formerly d9377b4647a43bd8d5c5bf1b3fb14b6d071dd135 [formerly 9c9cd39a0c9bc1f14e9dfc29ad98475794dda6b8]]
Former-commit-id: d4cb62bef897c4c33fd2a8ec6c909c58543d3eea
Former-commit-id: 302eec939ca95c8bc1fcfc53e8efbca9d4552eeb