t7201.24: Add refspec to keep --track working
We are formalizing a requirement that any remote-tracking branch to be used
as an upstream (i.e. as an argument to --track), _must_ "belong" to a
configured remote by being matched by the "dst" side of a fetch refspec.

Without this patch, this test would start failing when the new behavior is
introduced.

Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 530fed45ba3c046e954eaa62911313acfa196610 [formerly 530fed45ba3c046e954eaa62911313acfa196610 [formerly 88a9f72fe0f2ae3243b3d7a53dddf856cde48aca]]
Former-commit-id: 4c93958e2c90be2abc0eb4800c1a3dbfe8d2d0b6
Former-commit-id: ddcd882535133ca19a4a40b91a312e11f3c05da7