t9114.2: Don't use --track option against "svn-remote"-tracking branches
We are formalizing a requirement that any remote-tracking branch to be used
as an upstream (i.e. as an argument to --track), _must_ "belong" to a
configured remote by being matched by the "dst" side of a fetch refspec.

This test uses --track against a "remotes/trunk" ref which does not belong
to any configured (git) remotes, but is instead created by "git svn fetch"
operating on an svn-remote. It does not make sense to use an svn-remote as
an upstream for a local branch, as a regular "git pull" from (or "git push"
to) it would obviously fail (instead you would need to use "git svn" to
communicate with this remote). Furthermore, the usage of --track in this
case is unnecessary, since the upstreaming config that would be created is
never used.

Simply removing --track fixes the issue without changing the expected
behavior of the test.

Acked-by: Eric Wong <normalperson@yhbt.net>
Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 6a37e3d7eafcf3636675cdc9928de39311e7e2b8 [formerly 6a37e3d7eafcf3636675cdc9928de39311e7e2b8 [formerly 983b17d4bbedd426d4fdd8a95df4dcd292c1a695]]
Former-commit-id: 604604ad9e7b5d8c0b924eac8b2c9a75e5f64eeb
Former-commit-id: 9f77840ee6bf2a289057b227e063ed43e720dc5d