glossary: Update and rephrase the definition of a remote-tracking branch
The definition of a remote-tracking branch in the glossary have been
out-of-date for a while (by e.g. referring to "Pull:" from old-style
$GIT_DIR/remotes files).

Also, the preceding patches have formalized that a remote-tracking branch
must match a configured refspec in order to be usable as an upstream.

This patch rewrites the paragraph on remote-tracking branches accordingly.

Signed-off-by: Johan Herland <johan@herland.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fe526ea6e55cf14e16466c1edf0eeefa02df3c8d [formerly a5591cf587802e5af8f7a6b52f386ca951967621] [formerly a5591cf587802e5af8f7a6b52f386ca951967621 [formerly 229177aaea295a3f804f98494d7bcde9c088fc0a]]
Former-commit-id: 414687d42d06e1d840302f75aea27623c5a5a545
Former-commit-id: aab1e9d4f7c121eb631edabe465e3a9f23ba75d5