line-log: fix documentation formatting
The second paragraph of the added description for the -L option
"<start> and <end> can take one of these forms:", and the list of
forms that follow the headline, were indented one level too short,
due to the missing "+" to signal that the next paragraph continues
the previous one.

Also "You can specify this option more than once" is about the -L
option, not about its various forms of starting and ending points.
Move it to the end of the main text.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 1c86ee01739c64f5a2cafe396989f480686f7f15 [formerly 796dc117600ecb5cb9933e263365b3bb155f4a38] [formerly 796dc117600ecb5cb9933e263365b3bb155f4a38 [formerly 0f483436e9c2e03bbb80c1dc215d4d7bf8774282]]
Former-commit-id: 685146c9eb888573cdef10240969a1aec9393229
Former-commit-id: 41ade6ab82f4d949b22d704b7aba6ed7f1a8eb2f