git add: rephrase the "removal will cease to be ignored" warning
Now the logic to decide when to warn has been tightened, we know the
user is in a situation where the current and future behaviours will
be different.  Spell out what happens with these two versions and
how to explicitly ask for the behaviour, and suggest "git status" as
a way to inspect the current status.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 665f814e15e3acd6813b2aba9025d63d1f61ed83 [formerly 665f814e15e3acd6813b2aba9025d63d1f61ed83 [formerly ccc663bc24a3925b818df0ce97a5ba47e221f383]]
Former-commit-id: 0e9030aa1a82266dc52c92acbdb7ea878084fd85
Former-commit-id: 8301bded2bbbdc921fefce5cdb959a960a156082