git-shortlog.txt: make SYNOPSIS match log, update OPTIONS
There are broadly two problems with the current SYNOPSIS.  First, it
completely omits the detail that paths can be specified.  Second, it
attempts to list all the options: this is futile as, in addition to
the options unique to it, it accepts all the options that git-rev-list
accepts.  In fixing these problems, make the SYNOPSIS consistent with
that in git-log.txt.  Also add the corresponding sections to OPTIONS.
Save adding the options from rev-list-options.txt for a later patch,
as it requires some work to pick out the options that are relevant to
shortlog.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9fb9783ad798789fc44487e5c407754360c84f37 [formerly f4d01fba4c97cbe940376d6f6d9bfa73c8a2a8b7] [formerly f4d01fba4c97cbe940376d6f6d9bfa73c8a2a8b7 [formerly 46b2a46dd90f60b99faba8ee7c49c88cb58e0d8a]]
Former-commit-id: 47740e401901cc8052ef600492c5d49d4ef456df
Former-commit-id: e1a2055b066758e8e563dfec562857aae865ce6e