peel_ref(): fix return value for non-peelable, not-current reference
The old version was inconsistent: when a reference was
REF_KNOWS_PEELED but with a null peeled value, it returned non-zero
for the current reference but zero for other references.  Change the
behavior for non-current references to match that of current_ref,
which is what callers expect.  Document the behavior.

Current callers only call peel_ref() from within a for_each_ref-style
iteration and only for the current ref; therefore, the buggy code path
was never reached.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 02ef42616e7c21c34948f98efd51e5519c011367 [formerly 02ef42616e7c21c34948f98efd51e5519c011367 [formerly 2312a7932080f17c2847ec3ce5dddbc65c2e0b41]]
Former-commit-id: 52c661b80ec06c7d77619ac407ed8f20e17d8667
Former-commit-id: e00ef18e716356602a75d2bfde28e6d3f4792607