refs: extract a function peel_entry()
Peel the entry, and as a side effect store the peeled value in the
entry.  Use this function from two places in peel_ref(); a third
caller will be added soon.

Please note that this change can lead to ref_entries for unpacked refs
being peeled.  This has no practical benefit but is harmless.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e7a1ed90b73be1ab830b9687e056d3cac748cd11 [formerly e7a1ed90b73be1ab830b9687e056d3cac748cd11 [formerly 9a489f3c17d6c974b18c47cf406404ca2a721c87]]
Former-commit-id: 55ce0bac32df78aa2906d23072d93823c44fee10
Former-commit-id: 5fb00ace41960aa4f1dfd607c96a324df77fa1d9