t3210: test for spurious error messages for dangling packed refs
A packed reference can be overridden by a loose reference, in which
case the packed reference is obsolete and is never used.  The object
pointed to by such a reference can be garbage collected.  Since
a183cfacaf [formerly 1dce5f602e] [formerly 0eb226aaa8] [formerly f1b2cefd44 [formerly f1b2cefd44 [formerly d66da478f2]]], this could lead to the emission of a spurious error
message:

    error: refs/heads/master does not point to a valid object!

The error is generated by repack_without_ref() if there is an obsolete
dangling packed reference in packed-refs when the packed-refs file has
to be rewritten due to the deletion of another packed reference.  Add
a failing test demonstrating this problem and some passing tests of
related scenarios.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 52db38f83b805361fc755eca2819014fd51f46d1 [formerly 52db38f83b805361fc755eca2819014fd51f46d1 [formerly 0a0b8d1531feced259223314238aa43cb8ca2ce4]]
Former-commit-id: 0da0381242f2fa58634c4b14f2cadb84b8eec452
Former-commit-id: 385d5f307bec5baa10ab70fe4665531fe01c23d8