git add: --ignore-removal is a better named --no-all
In the historical context of "git add --all ." that pays attention
to "all kinds of changes" (implying "without ignoring removals"),
the option to countermand it "--no-all" may have made sense, but
because we will be making "--all" the default when a pathspec is
given, it makes more sense to rename the option to a more explicit
"--ignore-removal".  The "--all" option naturally becomes its
negation, "--no-ignore-removal".

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7fbc2725d27bc3145f3b0a5c10c458b7e0812035 [formerly 7fbc2725d27bc3145f3b0a5c10c458b7e0812035 [formerly 9f60f49b9222f9200e688a64cd80610305b03cc5]]
Former-commit-id: 1f3a0eb6e8e06fb7fda09e7ab8c76e3c557eeb7f
Former-commit-id: efd6cd5f68ccee9d50567665be355b04b65bb518