git add <pathspec>... defaults to "-A"
Make "git add <pathspec>..." notice paths that have been removed
from the working tree, i.e. the same as "git add -A <pathspec>...".

Given that "git add <pathspec>" is to update the index with the
state of the named part of the working tree as a whole, it makes it
more intuitive, and also makes it possible to simplify the advice we
give while marking the paths the user finished resolving conflicts
with.  We used to say "to record removal as a resolution, remove the
path from the working tree and say 'git rm'; for all other cases,
edit the path in the working tree and say 'git add'", but we can now
say "update the path in the working tree and say 'git add'" instead.

As promised, this merges the temporary update_files_in_cache() helper
function back to add_files_to_cache() function.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 113ef5541064dfcc772046c9211465a8ecd950bb [formerly 113ef5541064dfcc772046c9211465a8ecd950bb [formerly fdc97abd4a8b04781356dc561b4ba5a634f9854f]]
Former-commit-id: 088df4f6b5fce546e1c72e353f4600610bb11431
Former-commit-id: e97a6052097ff5662bdce1478ed2cf38dc5aec5f