gitk: On OSX, bring the gitk window to front
On OSX, Tcl/Tk application windows are created behind all
the applications down the stack of windows.  This is very
annoying, because once a gitk window appears, it's the
downmost window and switching to it is pain.

The patch is: if we are on OSX, use osascript to
bring the current Wish process window to front.

Signed-off-by: Tair Sabirgaliev <tair.sabirgaliev@gmail.com>
Thanks-to: Stefan Haller <lists@haller-berlin.de>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: d6dfb9ac031a206f8e3b655e352f33a1c8749d77 [formerly d6dfb9ac031a206f8e3b655e352f33a1c8749d77 [formerly 76bf6ff93e1766b8813f0cdbe9089080010ba646]]
Former-commit-id: a4afc2f26adeaa8fd52365ef8d40e0aefaa056c9
Former-commit-id: 580a867f4a3bb37542def6dc4ba4149adff73589