remote: 'show' and 'prune' can take more than one remote
The 'git remote show' and 'prune' subcommands are documented as taking
only a single remote name argument, but that is not the case; they
will simply iterate the action over all remotes given.  Update the
documentation and tests to match.

With the last user of the -f flag gone, we also remove the code
supporting it.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: fb3d88837ca80635f727137a97837f74e8df60c6 [formerly 307b21314b06e725fb414c9799aea5ce9f1087bd] [formerly 307b21314b06e725fb414c9799aea5ce9f1087bd [formerly b17dd3f9d6b1657b3fbfb40a863fc2dce5f54798]]
Former-commit-id: 5c352856125953ac524e9c34e2088afb9051b806
Former-commit-id: c10da3a05f3aed1dea4e51350e605f8d3c572f23