t5801: "VAR=VAL shell_func args" is forbidden
It is not a portable expectation that a single-shot environment
variable assignment works when calling a shell function, not a
command.

Set and export the variable before calling "test_must_fail git push"
instead.  This change would not hurt because this is the last
command in the subprocess and the environment will not seep through
to later tests without using a single-shot assignment.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 081f0e744eacf79f09c638e734154a475f0048ef [formerly 8c5684dd59aa35f93495f3ff0efc2b5f9bb0fe26] [formerly 8c5684dd59aa35f93495f3ff0efc2b5f9bb0fe26 [formerly 1afe6e404453586a1095ce0f3f7e899e62ab237f]]
Former-commit-id: f1d06c24caf37a79f9d2b9af862c4caf745947dc
Former-commit-id: c31d527593cdfc78942d5f9c43eb1bdafdcb7b38