pretty: simplify input line length calculation in pp_user_info()
Instead of searching for LF and NUL with two strchr() calls use a single
strchrnul() call.  We don't need to check if the returned pointer is NULL
because either we'll find the NUL at the end of line, or the caller
forgot to NUL-terminate the string and we'll overrun the buffer in any
case.  Also we don't need to pass LF or NUL to split_ident_line() as it
ignores it anyway.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 743a15f53dc1883983e59bb9c1ce59a2ee3f49bf [formerly 743a15f53dc1883983e59bb9c1ce59a2ee3f49bf [formerly 30e77bcb5008d1947c721fb084e21eeaa591c45e]]
Former-commit-id: c76751b37e9475690cc322dc11ddfcbb804fe3b4
Former-commit-id: 2053f8556140b03943955f177d78721b2c153987