t/Makefile: remove smoke test targets
Commit d24fbca (Remove Git's support for smoke testing - 2011-12-23)
removed the smoke test support from the test suite but it was
re-added by commit 342e9ef (Introduce a performance testing
framework - 2012-02-17).  This appears to be the result of a
mis-rebase, since re-adding the smoke testing infrastructure does
not relate to the subject of that commit.

The current 'smoke' target is broken since the 'harness' script it
uses no longer exists, so just reapply this section of commit d24fbca
and remove all of the smoke testing section in the makefile.

Signed-off-by: John Keeping <john@keeping.me.uk>
Acked-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 194d6182a25c70cacf923ad4d9624fb1dba01552 [formerly 194d6182a25c70cacf923ad4d9624fb1dba01552 [formerly 6a776acabf1f83846e03acb0ccc4d14d0be5ab69]]
Former-commit-id: 2eed61a52136606be53d1a1b3a608f4c79e88c67
Former-commit-id: e8d06f0fe98e2a29992cd234f7b4f1d22e9824bc