t/Makefile: fix result handling with TEST_OUTPUT_DIRECTORY
When TEST_OUTPUT_DIRECTORY is set, the test results will be generated in
"$TEST_OUTPUT_DIRECTORY/test-results", which may not be the same as
"test-results" in t/Makefile.  This causes the aggregate-results target
to fail as it finds no count files.

Fix this by introducing TEST_RESULTS_DIRECTORY and using it wherever
test-results is referenced.

Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 99dcfe843ebf9af19b6236e1d27eab7803afe88d [formerly 99dcfe843ebf9af19b6236e1d27eab7803afe88d [formerly 54bb9015c817728371b48e908257fdf6a078eb60]]
Former-commit-id: d066dba0d0c7bbb8881df5fcd7d36f8a054352c5
Former-commit-id: 65870733cf04f56739de8bc18728341623a9bdbb