git add: avoid "-u/-A without pathspec" warning on stat-dirty paths
In preparation for Git 2.0, "git add -u/-A" without pathspec checks
all the working tree (not limited to the current directory) and
issues a warning when it finds any path that we might add in Git
2.0, because that would mean the users' fingers need to be trained
to explicitly say "." if they want to keep the current behaviour.

However, the check was incomplete, because "git add" usually does
not refresh the index, considers a path that is stat-dirty but has
contents that is otherwise up-to-date in the index as "we might
add", and relies on that it is a no-op to add the same thing again
via the add_file_to_index() API (which also knows not to say "added"
in verbose mode when this happens).  We do not want to trigger the
warning for a path that is outside the current directory is merely
stat-dirty, as it won't be added in Git 2.0, either.

Signed-off-by: Junio C Hamano <gitster@pobox.com>
Acked-by: Jonathan Nieder <jrnieder@gmail.com>

Former-commit-id: 06ee9af2aaa643b6680e82c0634ad7f145584862 [formerly 06ee9af2aaa643b6680e82c0634ad7f145584862 [formerly 0df7b8e55c3b193f2b2f60f7cac1ebb2ce04cd9e]]
Former-commit-id: c38919de303a25dff2675294e1ff920245166042
Former-commit-id: 1166de2fccaba71765f49cdd5116d1882b3b5f98