git-gui: fix file name handling with non-empty prefix
Commit e3d06ca (git-gui: Detect full path when parsing arguments -
2012-10-02) fixed the handling of absolute paths passed to the browser
and blame subcommands by checking whether the file exists without the
prefix before prepending the prefix and checking again.  Since we have
chdir'd to the top level of the working tree before doing this, this
does not work if a file with the same name exists in a subdirectory and
at the top level (for example Makefile in git.git's t/ directory).

Instead of doing this, revert that patch and fix absolute path issue by
using "file join" to prepend the prefix to the supplied path.  This will
correctly handle absolute paths by skipping the prefix in that case.

Acked-by: Andrew Wong <andrew.kw.w@gmail.com>
Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>

Former-commit-id: bf498c82bba47b05e08dcd9582d8d2c794602172 [formerly 5544e0650e3e399c5ccb907012c9480e6a6a5d3a] [formerly 5544e0650e3e399c5ccb907012c9480e6a6a5d3a [formerly 2f38dd03fcefb7ad3a6db2703eeb6dd9b4fd34b6]]
Former-commit-id: e13cd394a596681ff36742d1acf60af38fbd2c0b
Former-commit-id: a642cbac73e984cee03e53b190b1e2b5746ac354