compat/regex/regexec.c: Fix some sparse warnings
Sparse issues an "Using plain integer as NULL pointer" warning along
with two "symbol was not declared. Should it be static?" type warnings
for the 'merge_state_with_log' and 'find_recover_state' functions.

In order to suppress the warnings, we replace the use of '0' as a null
pointer constant with NULL and add the static modifier to the function
definitions.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 8faeead64a0bc0a64cb652c9a21958be009afd1d [formerly 8faeead64a0bc0a64cb652c9a21958be009afd1d [formerly 5b62e6374aabb20297b22bdd72952ff582b2198b]]
Former-commit-id: acd5a7b9fb8744a6f60978c6ffe463c77b927d0d
Former-commit-id: dae95e92da5f2d00524291089f9a69d428fc759b