compat/fnmatch/fnmatch.c: Fix a sparse error
Sparse issues the following error and warnings:

    SP compat/fnmatch/fnmatch.c
.../fnmatch.c:144:17: warning: Using plain integer as NULL pointer
.../fnmatch.c:238:67: warning: Using plain integer as NULL pointer
.../fnmatch.c:303:40: error: too many arguments for function getenv

The error is caused by the extern declaration for the getenv function
not including the function prototype. Without the prototype, sparse
effectively sees the declaration as 'char *getenv(void)'. In order to
suppress the error, we simply include the function prototype.

In order to suppress the warnings, we include the <stddef.h> header
which provides an appropriate definition of the NULL macro, rather
than using the (inappropriate) default definition at fnmatch.c:132.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7af0f040b5ca4dc79f6fe7d4b0a322c68d33059f [formerly 1abc10e04f0ff984a9b3c785755b1f6cff191156] [formerly 1abc10e04f0ff984a9b3c785755b1f6cff191156 [formerly 4fc8fb48e978b8563493d04069c2507b7fb0064a]]
Former-commit-id: b741f237b0390840fe6e646ab81e32b96247b388
Former-commit-id: 2729b4c5473ae10cf3981f4583cd8caa7e65e1ac