compat/unsetenv.c: Fix a sparse warning
The gitunsetenv function includes an (redundant) declaration of the
'environ' symbol, which is a pointer to the table of environment
variables. Unfortunately, on MinGW, this provokes sparse to issue
the following warning:

    compat/unsetenv.c:5:20: warning: non-ANSI function declaration of \
    function '__p__environ'

On MinGW, the <stdlib.h> header defines the 'environ' symbol as a
preprocessor macro (via _environ) which obtains the environ table
pointer via a call to the __p__environ() function.

In order to suppress the warning, we simply remove the redundant
declaration of the 'environ' symbol, since the symbol is already
declared correctly in <stdlib.h> (included via git-compat-util.h).

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 413afa0b2716d8ca7ab6f71c8ff3cc9673b067d7 [formerly 80aa147634adc62b3125e3822eaf6ea25c75eaa7] [formerly 80aa147634adc62b3125e3822eaf6ea25c75eaa7 [formerly ec535cc27e6c4f5e0b1d157e04f5511f166ecd9d]]
Former-commit-id: 1516481d81244e6368e7be325039d631bdd77dab
Former-commit-id: 63d0e9daf9bbb15cbc408e092f0665b0a8ce1410