compat/poll/poll.c: Fix a sparse warning
Sparse issues an 'Using plain integer as NULL pointer' warning when
passing the constant '0' as the second parameter in the call to the
WSAEventSelect() function. The function parameter has a pointer type
(WSAEVENT, aka HANDLE, aka void *) so that, in order to suppress the
warning, we simply pass NULL for that parameter in the function call
expression.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: bd97ac7375da8daa4b3d73f726ecb01f036a4851 [formerly bd97ac7375da8daa4b3d73f726ecb01f036a4851 [formerly 1c31596a4b029e51cef1c520f32a3ecfbdea3c0a]]
Former-commit-id: c9ce21f8db676dcfad6513ce16b47f94ed601a40
Former-commit-id: 9c8339c6d32dca05aeffdc262289658edb3acf00