compat/win32mmap.c: Fix some sparse warnings
Sparse issues two 'Using plain integer as NULL pointer' warnings
against the call to the CreateFileMapping() function. The warnings
relate to the second and sixth parameters, which both have pointer
type. In order to suppress the warnings, we simply pass the NULL
pointer, rather than '0', to those parameters in the function call.

Signed-off-by: Ramsay Jones <ramsay@ramsay1.demon.co.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: ce8fb66f1fed4ff54678244c034beea6370a09ab [formerly 2362d1fae1fadc5d797e0c74697ee7868e643260] [formerly 2362d1fae1fadc5d797e0c74697ee7868e643260 [formerly 15b7f601fc976f7a112035fb0e91869eabe61c26]]
Former-commit-id: 9d29264c9b4071721b35d563d8ccbff4be2099dd
Former-commit-id: 8ea38cfe2709c8f1fa0d9af95508bc2d483ee94b