completion: refactor diff_index wrappers
At the end of the day what we really need is to find out the files that
have been staged, or modified, because those files are the ones that
make sense to pass as arguments to 'git commit'.

We need diff-index to find those out, since 'git ls-files' doesn't do
that.

But we don't need wrappers and wrappers basically identical to the ones
used for 'git ls-files', when we can pretend it receives a --committable
option that would return what we need.

That way, we can remove a bunch of code without any functional changes.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d4eb1c5559b2e6424c5f55618ea950e83f5c76ff [formerly d4eb1c5559b2e6424c5f55618ea950e83f5c76ff [formerly f825972c38ba5a8e0309a552b401ad55aeaaaf36]]
Former-commit-id: a80f9993ea84db768604ab63ac88266f60d195e9
Former-commit-id: 0111cb47c182f57a38b79074d88a7fefd53f53c5