complete: zsh: trivial simplification
There should be no functional changes.

The only reason I wrapped this code around a sub-function is because zsh
did the same in it's bashcompinit script in order to declare the special
variable 'words' as hidden, but only in this context.

There's no need for that any more since we access __git_main directly,
so 'words' is not modified, so there's no need for the sub-function.

In zsh mode the array indexes are different though.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d8ea4d9fa55e1c533e6f94ed0b2723b4042839bc [formerly d8ea4d9fa55e1c533e6f94ed0b2723b4042839bc [formerly 1ca6d4bc426435ecb87c6eb7802e83ff0a2f1667]]
Former-commit-id: 67b49878fb104c8478a890c147fa117fef91cfae
Former-commit-id: ab7f403d23c78c15b12e9f1ff39ac654d6920d34