gitk: Display the date of a tag in a human-friendly way
By selecting a tag within gitk you can display information about it.
This information is output by using the command

 'git cat-file tag <tagid>'

This outputs the *raw* information from the tag, amongst which is the
time - in seconds since the epoch. As useful as that value is, I find it
a lot easier to read and process time which it is something like:

 "Mon Dec 31 14:26:11 2012 -0800"

This change will modify the display of tags in gitk like so:

  @@ -1,7 +1,7 @@
   object 1335555a198efcc39942c41cf626b6379a69579f [formerly b8dd0f7fdc3f45e1e1c1b682bc20514e44047706] [formerly e4bcf6c35375a86261dabbf795afe9c7b969b99c] [formerly ec61d5dc3cbb1a1d521cfd7fddda6407032b231a [formerly ec61d5dc3cbb1a1d521cfd7fddda6407032b231a [formerly 5d417842efeafb6e109db7574196901c4e95d273]]]
   type commit
   tag v1.8.1
  -tagger Junio C Hamano <gitster@pobox.com> 1356992771 -0800
  +tagger Junio C Hamano <gitster@pobox.com> Mon Dec 31 14:26:11 2012 -0800

   Git 1.8.1
   -----BEGIN PGP SIGNATURE-----

Signed-off-by: Anand Kumria <wildfire@progsoc.org>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: a07d330695e66e5b57c5c2e7e7b46a1084894c70 [formerly a07d330695e66e5b57c5c2e7e7b46a1084894c70 [formerly 685316c419caf4f4ec2e81e607d3e3ab089e329c]]
Former-commit-id: 455e1281e2068a6cdf003de2ea82503c92f6c50a
Former-commit-id: 912f08c8905191b88e943c841bd9b48033a23b73