upload-pack: ignore 'shallow' lines with unknown obj-ids
When the client sends a 'shallow' line for an object that the server does
not have, the server currently dies with the error: "did not find object
for shallow <obj-id>".  The client may have truncated the history at
the commit by fetching shallowly from a different server, or the commit
may have been garbage collected by the server. In either case, this
unknown commit is not relevant for calculating the pack that is to be
sent and can be safely ignored, and it is not used when recomputing where
the updated history of the client is cauterised.

The documentation in technical/pack-protocol.txt has been updated to
remove the restriction that "Clients MUST NOT mention an obj-id which it
does not know exists on the server". This requirement is not realistic
because clients cannot know whether an object has been garbage collected
by the server.

Signed-off-by: Michael Heemskerk <mheemskerk@atlassian.com>
Reviewed-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 714a28e85ee7eaffe953ac37afde388f8a7ea456 [formerly 66c30c0ffe6fd66d7ccb01574b66ab89d44a895f] [formerly 66c30c0ffe6fd66d7ccb01574b66ab89d44a895f [formerly af04fa2a78af3f4dc07bdc3e3018ccf3baa88c20]]
Former-commit-id: 8eda6ae59a49f535452401cb9004750fac953e65
Former-commit-id: 762b6d99bded3022d63083c1d4d08a00c244416f