Git.pm: call tempfile from File::Temp as a regular function
We call File::Temp's "tempfile" function as a class method, but it was
never designed to be called this way. Older versions seemed to
tolerate it, but as of File::Temp 0.23, it blows up like this:

  $ git svn fetch
  'tempfile' can't be called as a method at .../Git.pm line 1117.

Fix it by calling it as a regular function, just inside the File::Temp
namespace.

Signed-off-by: H. Merijn Brand <h.m.brand@xs4all.nl>
Helped-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: a716e3b5b7b888da85f048177c1c77e947636144 [formerly a716e3b5b7b888da85f048177c1c77e947636144 [formerly eafc2dd59f4d28d4a01deb24df588fd7a29990d8]]
Former-commit-id: df65f988cd6981219c98d9adfecb2b18dfc29345
Former-commit-id: c6e50087f830b942f398aee0bc67fdace0129423