t0008: use named pipe (FIFO) to test check-ignore streaming
sleeps in the check-ignore test suite are not ideal since they can
fail when the system is under load, or when a tool like valgrind is
used which drastically alters the timing.  Therefore we replace them
with a more robust solution using a named pipe (FIFO).

Thanks to Jeff King for coming up with the redirection wizardry
required to make this work.

http://article.gmane.org/gmane.comp.version-control.git/220916

Signed-off-by: Adam Spiers <git@adamspiers.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 25d0174b2833b5083aa6b86f685bbcf0be5bfc46 [formerly 4ab6285b37068170ab5171c4686bcb35529d015a] [formerly 4ab6285b37068170ab5171c4686bcb35529d015a [formerly b96114edb36e16d214af1e7945a1d19b8a5e1686]]
Former-commit-id: 145e5ecb28bb118c2e603f6adadfaf9a009e0771
Former-commit-id: 834befaae43f54a72fbeb2228435aa53d309aa30