remote-bzr: add support to push merges
In order to do that, we need to store the marks of every file, so that
they can be fetched when needed. Unfortunately we can't tell bazaar that
nothing changed, we need to send the data so that it can figure it out
by itself.

And since it will be requesting a bunch of information by the file_id,
it's better to have a helper dict (rev_files), so that we can fetch it
quickly.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 06b41cdc7446c0863dc9a40777343802bda85779 [formerly 06b41cdc7446c0863dc9a40777343802bda85779 [formerly f38dfc4c32507de2f7cf05350fe8b34fd4edd45d]]
Former-commit-id: d7c1537cf1ef06ce17728596a2abe0d0ae1cee11
Former-commit-id: bdfd753d8e8d773eafdf9b6ad6624c26a176422e