remote-bzr: fix partially pushed merge
If part of the merge was already pushed, we don't have the blob_marks
available, however, the commits are already stored in bazaar, so we can
use the revision_tree to fetch the contents.

We want to do this only when there's no other option.

There's no easy way to test this.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 61195f45d02ff6b1c6f464f446c1a001ee8ba66d [formerly 61195f45d02ff6b1c6f464f446c1a001ee8ba66d [formerly b25df87fad2b12349a55146d5c87625109a58221]]
Former-commit-id: c972c980ac2dc8fdd01c72b6d4e7925f277b0abe
Former-commit-id: 5fbbf8ddf3ab5eb9c2339b4279340d108b9ec171