contrib/subtree: don't delete remote branches if split fails
When using "git subtree push" to split out a subtree and push it to a
remote repository, we do not detect if the split command fails which
causes the LHS of the refspec to be empty, deleting the remote branch.

Fix this by pulling the result of the split command into a variable so
that we can die if the command fails.

Reported-by: Steffen Jaeckel <steffen.jaeckel@stzedn.de>
Signed-off-by: John Keeping <john@keeping.me.uk>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7ee09399eb61c1677d020f6437d3dea01ad8e654 [formerly 91a0d200a815234fcd0309747c692ad455c49081] [formerly 91a0d200a815234fcd0309747c692ad455c49081 [formerly 3212d56ce528a86170f7809659f80a8769a5875f]]
Former-commit-id: 58be8f7727477594f23ff363dfa1c1146fa1dc61
Former-commit-id: ef3763cfd19372137bc157ab99379c2bfdedf142