git-svn: added an --include-path flag
The SVN::Fetcher module is now able to filter for inclusion as well
as exclusion (as used by --ignore-path). Also added tests, documentation
changes and git completion script.

If you have an SVN repository with many top level directories and you
only want a git-svn clone of some of them then using --ignore-path is
difficult as it requires a very long regexp. In this case it's much
easier to filter for inclusion.

[ew: remove trailing whitespace]

Signed-off-by: Paul Walmsley <pjwhams@gmail.com>
Signed-off-by: Eric Wong <normalperson@yhbt.net>

Former-commit-id: 9d73aa3504dea5dfbbdf16428bfdb3b952cf888f [formerly 8ce53498b49be9a9f8363fb3c6f9b264e8f67815] [formerly 8ce53498b49be9a9f8363fb3c6f9b264e8f67815 [formerly a7b102302ad9495fb4eb5088ffcb09c44a406c06]]
Former-commit-id: 3cbf776d0ea1d10da865732c35aab001f6205eca
Former-commit-id: 67c7de09d76d9c5ce7b08bca5ffc6c07e47f5024