help: add help_unknown_ref()
When the user gives an unknown string to a command that expects to
get a ref, we could be more helpful than just saying "that's not a
ref" and die.

Add helper function help_unknown_ref() to take care of displaying an
error message along with a list of suggested refs the user might
have meant.  An interaction with "git merge" might go like this:

	$ git merge foo
	merge: foo - not something we can merge

	Did you mean one of these?
	    origin/foo
	    upstream/foo

Signed-off-by: Vikrant Varma <vikrant.varma94@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 382de19b62c1bc456596ae945e06e3fa98a0bd53 [formerly f40e3ffc22655188c6acee3c971604db07a935e9] [formerly f40e3ffc22655188c6acee3c971604db07a935e9 [formerly e56181060e1117d458b854c34e63206117dab8b6]]
Former-commit-id: 5c1254556e1b3dd3afa8b9c2d4d060f1cf287626
Former-commit-id: 2285eb394ecc427d88f794ab4cb849cc0710f882