remote-bzr: avoid bad refs
Versions of fast-export before v1.8.2 throws a bad 'reset' commands
because of a behavior in transport-helper that is not even needed.
We should ignore them, otherwise we will treat them as branches and
fail.

This was fixed in v1.8.2, but some people use this script in older
versions of git.

Also, check if the ref was a tag, and skip it for now.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 258f952addd4dab1e1ae9d1716cfb0d50813c8af [formerly bd794b142c0fa1849a4634b1cd0fd7d826caaf1b] [formerly bd794b142c0fa1849a4634b1cd0fd7d826caaf1b [formerly 4c00819910706aa8ae450292085e47e52c333440]]
Former-commit-id: d5dcd3e3e37118ab3e442e30fa0e5ba300491697
Former-commit-id: fff6dac2fe671a3f850842b1843f5fd363b3d631