fast-export: do not parse non-commit objects while reading marks file
We read from the marks file and keep only marked commits, but in
order to find the type of object, we are parsing the whole thing,
which is slow, specially in big repositories with lots of big files.

There's no need for that, we can query the object information with
sha1_object_info().

Before this, loading the objects of a fresh emacs import, with 260598
blobs took 14 minutes, after this patch, it takes 3 seconds.

This is the way fast-import does it. Also die if the object is not
found (like fast-import).

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 55ee019b9fe08ced9032a0fb573cbf65efb1a0d1 [formerly c7175efa2b638b81db357a255b897684d886ab1a] [formerly c7175efa2b638b81db357a255b897684d886ab1a [formerly e6812cfa9aba69a8c9d83b0710291b27bff0f7a3]]
Former-commit-id: e8e7006d9ca89c214c53abfd0b0b30d8f64d7944
Former-commit-id: a44fd088806feaa9e63ad2d52cc47bdc282cbd4b