merge-tree: handle directory/empty conflict correctly
git-merge-tree causes a null pointer dereference when a directory
entry exists in only one or two of the three trees being compared with
no corresponding entry in the other tree(s).

When this happens, we want to handle the entry as a directory and not
attempt to mark it as a file merge.  Do this by setting the entries bit
in the directory mask when the entry is missing or when it is a
directory, only performing the file comparison when we know that a file
entry exists.

Reported-by: Andreas Jacobsen <andreas@andreasjacobsen.com>
Signed-off-by: John Keeping <john@keeping.me.uk>
Tested-by: Andreas Jacobsen <andreas@andreasjacobsen.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d571f0162279f33ae46305117793e53c59450b6d [formerly d571f0162279f33ae46305117793e53c59450b6d [formerly 94883b4302118fe6ea50028d02bb453c9af38b66]]
Former-commit-id: 249c5819ec8e4e1161425858c62211c128a8ad0a
Former-commit-id: d1e2b087ba9417406dc5889b07bf3799dc5cd0f8