tests: at-combinations: check ref names directly
Some committishes might point to the same commit, but through a
different ref, that's why it's better to check directly for the ref,
rather than the commit message.

We can do that by calling rev-parse --symbolic-full-name, and to
differentiate the old from the new behavior we add an extra argument to
the check() helper.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0ddd69fcbfc10aae81ee3c86f9d10f06eec63444 [formerly 301d37d68bb447719d6a733dcf61cc3c469a2b7d] [formerly 301d37d68bb447719d6a733dcf61cc3c469a2b7d [formerly c8a81e90acda9da55bae627928e80b04e98980fc]]
Former-commit-id: 0ff7cd9972feee9523046128cd2922f50cfc154d
Former-commit-id: 40ccc60ee2e5e5f9893fdc9bd4daffd4226129f5