sha1_name: reorganize get_sha1_basic()
Through the years the functionality to handle @{-N} and @{u} has moved
around the code, and as a result, code that once made sense, doesn't any
more.

There is no need to call this function recursively with the branch of
@{-N} substituted because dwim_{ref,log} already replaces it.

However, there's one corner-case where @{-N} resolves to a detached
HEAD, in which case we wouldn't get any ref back.

So we parse the nth-prior manually, and deal with it depending on
whether it's a SHA-1, or a ref.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d8df7dc04d1eee6a0db93e835bf26acf2e18f39b [formerly d8df7dc04d1eee6a0db93e835bf26acf2e18f39b [formerly 128fd54daeff23fbef0d0b7ec1a31f38d0803f63]]
Former-commit-id: e962c1cb45818b9ed0767a1fcffc75888d3ddca3
Former-commit-id: da3f24d0d58af74178fd211012bfcc092a488391