Add new @ shortcut for HEAD
Typing 'HEAD' is tedious, especially when we can use '@' instead.

The reason for choosing '@' is that it follows naturally from the
ref@op syntax (e.g. HEAD@{u}), except we have no ref, and no
operation, and when we don't have those, it makes sens to assume
'HEAD'.

So now we can use 'git show @~1', and all that goody goodness.

Until now '@' was a valid name, but it conflicts with this idea, so
let's make it invalid. Probably very few people, if any, used this name.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 777f8e94e04200cc72887c5f8f898cdd63db1be8 [formerly 777f8e94e04200cc72887c5f8f898cdd63db1be8 [formerly cdfd94837b27c220f70f032b596ea993d195488f]]
Former-commit-id: 5843602db9a7bb328ff383ce9680e16005df0c89
Former-commit-id: 9e0f381a2c96837c1a18429edc6a519fb04291de