t5004: ignore pax global header file
Versions of tar that don't know pax headers -- like the ones in NetBSD 6
and OpenBSD 5.2 -- extract them as regular files.  Explicitly ignore the
file created for our global header when checking the list of extracted
files, as this is normal and harmless fall-back behaviour.  This fixes
test 3 of t5004 on these platforms.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 40f73fdf8869df0c97c846b8d9a6e4a601a947e4 [formerly 40f73fdf8869df0c97c846b8d9a6e4a601a947e4 [formerly abdb9b2e4fd7876ab7138119c4aedca70bc0d250]]
Former-commit-id: 9d3f62daf7df4c2fd0e9d0268e10bdeb485a1eba
Former-commit-id: c450e682ea412a181f97d896c16b81dfb6a8d61e