t5004: resurrect original empty tar archive test
Add a test to verify the emptiness of an archive by extracting its
contents.  Don't run this test if the version of tar doesn't support
archives containing only a comment header, though.

The existing check 'tar archive of empty tree is empty' used to work
like that (minus the tar capability check) but was changed to depend
on the exact representation of empty tar files created by git archive
instead of on the behaviour of tar in order to avoid issues with
different tar versions.

The different approaches test different things: The existing one is
for empty trees, for which we know the exact expected output and thus
we can simply check it without extracting; the new one is for commits
with empty trees, whose archives include stamps and so the more
"natural" check by extraction is a better fit because it focuses on
the interesting aspect, namely the absence of any archive entries.

Signed-off-by: Rene Scharfe <rene.scharfe@lsrfire.ath.cx>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 4747bec152fe4f915a1bb38e8d614cbb02fcff5e [formerly 4747bec152fe4f915a1bb38e8d614cbb02fcff5e [formerly 56ee96572a7ec5a6b1a59ccdbbf204af16903d17]]
Former-commit-id: c5ad3fbb23fb7941907fda7c59ac6097b3702557
Former-commit-id: 3f3d49dffde50ae52a4d0312561e5ddb12fe5395