show: honor --textconv for blobs
Currently, "diff" and "cat-file" for blobs honor "--textconv" options
(with the former defaulting to "--textconv" and the latter to
"--no-textconv") whereas "show" does not honor this option, even though
it takes diff options.

Make "show" on blobs honor "--textconv" when it is asked.  The default
is not to apply textconv, which is in line with what "cat-file" does.

Signed-off-by: Michael J Gruber <git@drmicha.warpmail.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e892e310e571e02b0e0a2d92cd4366b5d3860263 [formerly e892e310e571e02b0e0a2d92cd4366b5d3860263 [formerly 083b99310978c6768d8e954c9bab1d3692a515a1]]
Former-commit-id: 6fbd08f80405e8f361477f9c656f515417cb95df
Former-commit-id: 31a607924bd3470108e0ff20f978d5b572201cd7