compate/clipped-write.c: large write(2) fails on Mac OS X/XNU
Due to a bug in the Darwin kernel, write(2) calls have a maximum size
of INT_MAX bytes.

Introduce a new compat function, clipped_write(), that only writes
at most INT_MAX bytes and returns the number of bytes written, as
a substitute for write(2), and allow platforms that need this to
enable it from the build mechanism with NEEDS_CLIPPED_WRITE.

Set it for Mac OS X by default.  It may be necessary to include this
function on Windows, too.

Signed-off-by: Filipe Cabecinhas <filcab+git@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b35cc0ebf19f012cb8daccb960de88a25998d5ae [formerly dbeaff7a83a3fa15ffe65b4ef6aeefaee99ed0e4] [formerly dbeaff7a83a3fa15ffe65b4ef6aeefaee99ed0e4 [formerly 6c642a878688adf46b226903858b53e2d31ac5c3]]
Former-commit-id: aa3067e22d68aeb53e20fdb1f6e245a87f724bf3
Former-commit-id: af02c74c642491de67646ddab6403e52650b2a50