gitk: Improve behaviour of drop-down lists
The drop-down lists used for things like the criteria for finding
commits (containing/touching paths/etc.) use a combobox if we are
using the ttk widgets.  By default the combobox exports its value
as the selection when it is changed, which is unnecessary, and sometimes
the combobox wouldn't release the selection, which is annoying.

To fix this, we make these comboboxes not export their selection,
and also clear their selection whenever they are changed.  This makes
them more like a simple selection of alternatives, improving the look
and feel of gitk.

Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: 29cd6599c83940b6a2ef882867f28f6f7141b703 [formerly 7f82fe6d22cb6a8afd15f198cf505ebf1cf30da6] [formerly 7f82fe6d22cb6a8afd15f198cf505ebf1cf30da6 [formerly 39c126914b4b7365251f32d34926506c229f090e]]
Former-commit-id: d04cac0848b821b1e2c6cefdce943ddd9cca5503
Former-commit-id: 008e5d88e2877dbab12a0c9e34b74002c79d08e3