refactor "ref->merge" flag
Each "struct ref" has a boolean flag that is set by the
fetch code to determine whether the ref should be marked as
"not-for-merge" or not when we write it out to FETCH_HEAD.

It would be useful to turn this boolean into a tri-state,
with the third state meaning "do not bother writing it out
to FETCH_HEAD at all". That would let us add extra refs to
the set of refs to be stored (e.g., to store copies of
things we fetched) without impacting FETCH_HEAD.

This patch turns it into an enum that covers the tri-state
case, and hopefully makes the code more explicit and easier
to read.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0665d733a8b1bbb9a52014c7f52917794951f7ad [formerly 8530e695a6474680b5facc8b4b2190408d8fc9ca] [formerly 8530e695a6474680b5facc8b4b2190408d8fc9ca [formerly 900f2814b8da951a450f5762dabb1b248dd77abc]]
Former-commit-id: 4cda0e8a659ba0b24cb60186ecc110d92d45c2eb
Former-commit-id: cdc68348b4857bb74f8ca8a18e09e4e928f3a4ef