fetch: opportunistically update tracking refs
When we run a regular "git fetch" without arguments, we
update the tracking refs according to the configured
refspec. However, when we run "git fetch origin master" (or
"git pull origin master"), we do not look at the configured
refspecs at all, and just update FETCH_HEAD.

We miss an opportunity to update "refs/remotes/origin/master"
(or whatever the user has configured). Some users find this
confusing, because they would want to do further comparisons
against the old state of the remote master, like:

  $ git pull origin master
  $ git log HEAD...origin/master

In the currnet code, they are comparing against whatever
commit happened to be in origin/master from the last time
they did a complete "git fetch".  This patch will update a
ref from the RHS of a configured refspec whenever we happen
to be fetching its LHS. That makes the case above work.

The downside is that any users who really care about whether
and when their tracking branches are updated may be
surprised.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 27c57165072afb93741fbbf6b35e53fe3dc4749c [formerly 5313e93dd56700672744e45c3caab4e220a08ec0] [formerly 5313e93dd56700672744e45c3caab4e220a08ec0 [formerly f269048754f3b835f4f7287c5a132714a059efce]]
Former-commit-id: 9aaddb7b3f0801c3929ae4c9f98e2af85cbb1335
Former-commit-id: 728841dd7612420ed78e0a081d9255b5abe07660