rebase: prepare to do generic housekeeping
On successful completion of a rebase in git-rebase--$backend.sh, the
$backend script cleans up on its own and exits.  The cleanup routine
is however, independent of the $backend, and each $backend script
unnecessarily duplicates this work:

    rm -rf "$state_dir"
    git gc --auto

Prepare git-rebase.sh for later patches that return control from each
$backend script back to us, for performing this generic cleanup
routine.  The code that this patch adds is currently unreachable, and
will only start to be used when git-rebase--$backend.sh scripts are
taught to return control in later patches.

Another advantage is that git-rebase.sh can implement a generic
finish_rebase() to possibly do additional tasks in addition to the
cleanup.

Signed-off-by: Ramkumar Ramachandra <artagnon@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0505a046f358b1d067ad675bf2c7172a58b01f5d [formerly 0505a046f358b1d067ad675bf2c7172a58b01f5d [formerly f5f758a5df470352ed46863288551f25fe9c7ea7]]
Former-commit-id: c6fdd3dffbe019b332ce3c8990bf7c213c67a431
Former-commit-id: 4b6366852923669d5f7ed4d1ab709f05163b16a3