t5551: do not use unportable sed '\+'
The set-up step to prepare a repository with 50000 tags used a
non-porable '\+' to match one-or-more.

The error was not caught because the next test that uses that
repository did not even bother to check if these expected tags were
actually cloned to the resulting repository.

Fix the sed construct to use BRE and update the "clone" test that
wanted to test cloning from such a repository with many refs to
check the resulting repository.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 5e060bc4b630be7cf0bab9b6df0ba7739a511493 [formerly c44bbdc6fff6d8e7f3d8fe5526f33bd6cc7e422f] [formerly c44bbdc6fff6d8e7f3d8fe5526f33bd6cc7e422f [formerly 5e2c7cd2c124da05d1807d5ffe61b2771bb755ee]]
Former-commit-id: 0c15f2f6f699d9e34996dde94ea5d50a4f9bc755
Former-commit-id: 4d8184185c4680db64f40857cbe6c4a203e0a157