coverage: do not delete .gcno files before building
The coverage-compile target depends on coverage-clean, which is
supposed to remove the earlier build products that would get in the
way of the next coverage test run.

However, removing *.gcno is actively wrong.  These are the files that
contain the compile-time coverage related data.  They are only rebuilt
if the source is compiled.  So if one ran 'make coverage' two times in
a row, the second run would remove *.gcno, but then fail to recreate
them because neither source files nor build flags have changed.  (This
remained hidden for so long most likely because any other intervening
use of 'make' will change the build flags, causing a full rebuild.)

So we make an exception for *.gcno.  The *.gcda are the coverage
results, written when the gcov-instrumented program is run.  We still
remove those, so as to get a one-test-run view of the data; you could
probably argue the other way too.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 19cd4fcab955860d4e1c0c852397d514e293ab02 [formerly bdc2ee018fe2633edbb35c0adaf303a3f0be34a8] [formerly bdc2ee018fe2633edbb35c0adaf303a3f0be34a8 [formerly dcbe7f1ab8b96b3875ae820d71c1364ee3bd99f5]]
Former-commit-id: b3e71db8d41abb630ca4de5e88c379ea7e3066de
Former-commit-id: 05e03587040530b84c13f419593ef58d17bb23d0