coverage: set DEFAULT_TEST_TARGET to avoid using prove
If the user sets DEFAULT_TEST_TARGET=prove in his config.mak, that
carries over into the coverage tests.  Which is really bad if he also
sets GIT_PROVE_OPTS=-j<..> as that completely breaks the coverage
runs.

Instead of attempting to mess with the GIT_PROVE_OPTS, just force the
test target to 'test' so that we run under make, like we intended all
along.

Signed-off-by: Thomas Rast <trast@inf.ethz.ch>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 43ae35ebd034d3f8981e3b76154237bf880f23c4 [formerly 43ae35ebd034d3f8981e3b76154237bf880f23c4 [formerly c14cc77c1126c04b35ce0ea16989af50af6cebc1]]
Former-commit-id: 71ebb800e134152ff936a2bdedd65f7c220e18a1
Former-commit-id: 67f60b1af1a4cb78f10e3df509fd739fc46bdd95