remote-hg: fix new branch creation
When a user creates a new branch with git:

 % git checkout -b branches/devel

and then pushes this branch

 % git push origin branches/devel

which is the way to push new mercurial branches, we do want to
create a branch, but the command would fail without newbranch=True.

This only matters when force_push=False, but setting newbranch=True
unconditionally does not hurt.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: e47c427a6be10ece30e6745e128ad966f7c9afa3 [formerly e47c427a6be10ece30e6745e128ad966f7c9afa3 [formerly 637333673a156898087e87f551a6eb1f9405e39d]]
Former-commit-id: ddce12d7ab1d29705bbb221a05ed3a1feea0bd0a
Former-commit-id: 397b0f4638e7e225658177be1a2e785d99425203