remote-hg: disable forced push by default
In certain situations we might end up pushing garbage revisions
(e.g. in a rebase), and the patches to deal with that haven't been
merged yet.  So let's disable forced pushes by default.

We are essentially reverting back to the old v1.8.2 behavior, to
minimize the possibility of regressions, but in a way the user can
configure.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 7d5680f6471a6f0c822bc9b4f042230af7e5dd11 [formerly 3facb2918e5980817782e18e999021b0d0a6a153] [formerly 3facb2918e5980817782e18e999021b0d0a6a153 [formerly 06f42133550725d360c9b6acc3b3fa223fee6c47]]
Former-commit-id: 61a4ecbcf6fe08e2bfb7e9e2cee3c291c8f8a948
Former-commit-id: 1bbcd3cfd32f85fb33c9771b4691eb9e03b98fc9