config: refactor management of color.ui's default value
The meaning of get_colorbool_found and get_diff_color_found is "the
config value if found, and -1 otherwise", but get_color_ui_found had a
slightly different meaning, as it has the value 0 (which corresponds to
the default value from the user point of view) when color.ui is unset.

Make get_color_ui_found default to -1, and make it explicit that 0 is the
default value when nothing else is found.

Signed-off-by: Matthieu Moy <Matthieu.Moy@imag.fr>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 0ea76fcc1e39debd7d43242b147b604029d8541f [formerly 0ea76fcc1e39debd7d43242b147b604029d8541f [formerly b8612b4da45087d256a0ce339bc65611fb6586f5]]
Former-commit-id: 7cd6ab787841311e2b01c243a1b90f26fe1a6130
Former-commit-id: 5b5d145042e8bdca4329015531b07dddb4c77419