strbuf_branchname(): do not double-expand @{-1}~22
If you were on 'frotz' branch before you checked out your current
branch, "git merge @{-1}~22" means the same as "git merge frotz~22".

The strbuf_branchname() function, when interpret_branch_name() gives
up resolving "@{-1}~22" fully, returns "frotz" and tells the caller
that it only resolved "@{-1}" part of the input, mistakes this as a
total failure, and appends the whole thing to the result, yielding
"frotz@{-1}~22", which does not make any sense.

Inspect the return value from interpret_branch_name() a bit more
carefully.  When it errored out without consuming anything, we will
get -1 and we should return the whole thing.  Otherwise, we should
append the remainder (i.e. "~22" in the earlier example) to the
partially resolved name (i.e. "frotz").

The test suite adds enough number of checkout to make @{-12} in the
last test in t0100 that tried to check "we haven't flipped branches
that many times" error case succeed; raise the number to a hundred.

Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: b7321af2a85bac03b6fc035f35077878eeca9bd2 [formerly affcfb27429381e73caa056c0595c3dce383f434] [formerly affcfb27429381e73caa056c0595c3dce383f434 [formerly 84cf246670eab56a23ed5554ed084053a0f19f2d]]
Former-commit-id: 5c29ab2fbd5f4faa3172ade24877c464a9855f97
Former-commit-id: 87f37b442b53a56def9069b06bdc4224a6b69770