t6019: test file dropped in -s ours merge
In preparation for upcoming TREESAME work, check the result for G.t,
which is dropped in "-s ours" merge L. The default rev-list is empty, as
expected - it follows the first parent path where it never existed.

Unfortunately, --ancestry-path is also empty. Merges H J and L are all
TREESAME to 1 parent, so are treated as TREESAME and not shown. This is
clearly undesirable in the case of merge L, which dropped our G.t by
taking the non-ancestry-path version. Document this as a known failure,
and expect "H J L", the 3 merges along the path that had to chose G.t
versions.

Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: d0e56358c04c06d90afcb6035c5893f0713b0af6 [formerly d0e56358c04c06d90afcb6035c5893f0713b0af6 [formerly c72424b1b5d511a346742dba62bec93a719ef1c8]]
Former-commit-id: ecf3b08d77b6b390bf4dad7255661c7e2586cb1f
Former-commit-id: 98d53387a602ded83d097f30907cf0e9c07d2832