simplify-merges: never remove all TREESAME parents
When simplifying an odd merge, such as one that used "-s ours", we may
find ourselves TREESAME to apparently redundant parents. Prevent
simplify_merges() from removing every TREESAME parent; if this would
happen reinstate the first TREESAME parent - the one that the default
log would have followed.

This avoids producing a totally disjoint history from the default log
when the default log is a better explanation of the end result, and aids
visualisation of odd merges.

Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 9997970b08467a4ed1ebbfcc0da5c393bd8c204c [formerly a0cd3efa5873d4474866aa1a565c9d8580f3ab8b] [formerly a0cd3efa5873d4474866aa1a565c9d8580f3ab8b [formerly 9c129eab99f16d7bb174b35573e68ae6a5d02efe]]
Former-commit-id: bec207c4be4fb2c974aff191d3a855c994c739b4
Former-commit-id: c33e0f8d2f4926f4942bc2c253f05b7463acd1b9