simplify-merges: drop merge from irrelevant side branch
Reimplement commit 4b7f53da on top of the new simplify-merges
infrastructure, tightening the condition to only consider root parents;
the original version incorrectly dropped parents that were TREESAME to
anything.

Original log message follows.

The merge simplification rule stated in 6546b59 (revision traversal:
show full history with merge simplification, 2008-07-31) still
treated merge commits too specially.  Namely, in a history with this
shape:

	---o---o---M
	          /
         x---x---x

where three 'x' were on a history completely unrelated to the main
history 'o' and do not touch any of the paths we are following, we
still said that after simplifying all of the parents of M, 'x'
(which is the leftmost 'x' that rightmost 'x simplifies down to) and
'o' (which would be the last commit on the main history that touches
the paths we are following) are independent from each other, and
both need to be kept.

That is incorrect; when the side branch 'x' never touches the paths,
it should be removed to allow M to simplify down to the last commit
on the main history that touches the paths.

Suggested-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c48e415a331ee5f8d197ec9f4b9e59a80ce0e960 [formerly 110f23cd8809124ebfcc63182b3c2f3cc968e3f2] [formerly 110f23cd8809124ebfcc63182b3c2f3cc968e3f2 [formerly 143f1eafdb0d326b1a8e3d15bf59a497b9440bf1]]
Former-commit-id: 0a7432135e134cd2936f6cc36c2275695b30878d
Former-commit-id: c181c08741ba53b253b83d0457f82e30ba1006c6