revision.c: discount side branches when computing TREESAME
Use the BOTTOM flag to define relevance for pruning. Relevant commits
are those that are !UNINTERESTING or BOTTOM, and this allows us to
identify irrelevant side branches (UNINTERESTING && !BOTTOM).

If a merge has relevant parents, and it is TREESAME to them, then do not
let irrelevant parents cause the merge to be treated as !TREESAME.

When considering simplification, don't always include all merges -
merges with exactly one relevant parent can be simplified, if TREESAME
according to the above rule.

These two changes greatly increase simplification in limited, pruned
revision lists.

Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: c21cb4ca95146089989d55784e71a941990eb64d [formerly c21cb4ca95146089989d55784e71a941990eb64d [formerly 4d826608e9610851d29440ca290e54b701921608]]
Former-commit-id: 76b618da4fe9f833e9f6569ca0da422461d7963c
Former-commit-id: 82937b7eef6948a3308dd1b6e012f09cce611c21