revision.c: don't show all merges for --parents
When using --parents or --children, get_commit_action() previously showed
all merges, even if TREESAME to both parents.

This was intended to tie together the topology of the rewritten parents,
but it was excessive - in fact we only need to show merges that have two
or more relevant parents. Merges at the boundary do not necessarily need
to be shown.

Signed-off-by: Kevin Bracey <kevin@bracey.fi>
Signed-off-by: Junio C Hamano <gitster@pobox.com>

Former-commit-id: 81e1814941993486c44b12e3f9c70cffcbc1f310 [formerly 81e1814941993486c44b12e3f9c70cffcbc1f310 [formerly bf3418b08bac89902a5f6c70f8695f148df99828]]
Former-commit-id: 28dfda5acccbd2f9e57b1bafbfb715247840dc11
Former-commit-id: 3dee295c8e917b748fab4fe1449af1ebc9dd4951